spil exec serial and parallel into own files
