+++ execution/stagedsync/exec3.go
@@ -22,6 +22,7 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
+	"math/big"
 	"slices"
 	"sync"
 	"sync/atomic"
@@ -37,22 +38,19 @@ import (
 	"github.com/erigontech/erigon/core/state"
 	"github.com/erigontech/erigon/core/tracing"
 	"github.com/erigontech/erigon/db/kv"
-	"github.com/erigontech/erigon/db/kv/mdbx"
 	"github.com/erigontech/erigon/db/kv/rawdbv3"
-	"github.com/erigontech/erigon/db/kv/temporal"
 	"github.com/erigontech/erigon/db/rawdb"
 	"github.com/erigontech/erigon/db/rawdb/rawdbhelpers"
 	"github.com/erigontech/erigon/db/rawdb/rawtemporaldb"
 	dbstate "github.com/erigontech/erigon/db/state"
-	"github.com/erigontech/erigon/db/state/changeset"
 	"github.com/erigontech/erigon/db/wrap"
 	"github.com/erigontech/erigon/execution/commitment"
 	"github.com/erigontech/erigon/execution/commitment/commitmentdb"
-	"github.com/erigontech/erigon/execution/consensus"
 	"github.com/erigontech/erigon/execution/exec3"
 	"github.com/erigontech/erigon/execution/types"
 	"github.com/erigontech/erigon/execution/types/accounts"
 	"github.com/erigontech/erigon/turbo/shards"
+	"golang.org/x/sync/errgroup"
 )
 
 // Cases:
@@ -131,10 +129,6 @@ func ExecV3(ctx context.Context,
 ) (execErr error) {
 	inMemExec := txc.Doms != nil
 
-	blockReader := cfg.blockReader
-	chainConfig := cfg.chainConfig
-	totalGasUsed := uint64(0)
-
 	useExternalTx := txc.Tx != nil
 	var applyTx kv.TemporalRwTx
 
@@ -192,7 +186,6 @@ func ExecV3(ctx context.Context,
 
 	var (
 		stageProgress = execStage.BlockNumber
-		outputTxNum   = atomic.Uint64{}
 		blockNum      = doms.BlockNum()
 	)
 
@@ -200,8 +193,7 @@ func ExecV3(ctx context.Context,
 		return nil
 	}
 
-	outputTxNum.Store(doms.TxNum())
-	agg.BuildFilesInBackground(outputTxNum.Load())
+	agg.BuildFilesInBackground(doms.TxNum())
 
 	var (
 		inputTxNum               uint64
@@ -241,74 +233,15 @@ func ExecV3(ctx context.Context,
 	logInterval := 20 * time.Second
 	logEvery := time.NewTicker(logInterval)
 	defer logEvery.Stop()
-	flushEvery := time.NewTicker(2 * time.Second)
-	defer flushEvery.Stop()
 	defer resetExecGauges(ctx)
 	defer resetCommitmentGauges(ctx)
 	defer resetDomainGauges(ctx)
 
-	var executor executor
-	var executorContext context.Context
-	var executorCancel context.CancelFunc
-
 	if parallel {
-		pe := &parallelExecutor{
-			txExecutor: txExecutor{
-				cfg:                   cfg,
-				rs:                    rs,
-				doms:                  doms,
-				agg:                   agg,
-				isMining:              isMining,
-				inMemExec:             inMemExec,
-				logger:                logger,
-				logPrefix:             execStage.LogPrefix(),
-				progress:              NewProgress(blockNum, outputTxNum.Load(), commitThreshold, false, execStage.LogPrefix(), logger),
-				enableChaosMonkey:     execStage.CurrentSyncCycle.IsInitialCycle,
-				hooks:                 hooks,
-				lastCommittedTxNum:    doms.TxNum(),
-				lastCommittedBlockNum: blockNum,
-			},
-			workerCount: workerCount,
-		}
-
-		executorContext, executorCancel = pe.run(ctx)
-
-		defer executorCancel()
-
-		executor = pe
-	} else {
-		se := &serialExecutor{
-			txExecutor: txExecutor{
-				cfg:                   cfg,
-				rs:                    rs,
-				doms:                  doms,
-				agg:                   agg,
-				u:                     u,
-				isMining:              isMining,
-				inMemExec:             inMemExec,
-				applyTx:               applyTx,
-				logger:                logger,
-				logPrefix:             execStage.LogPrefix(),
-				progress:              NewProgress(blockNum, outputTxNum.Load(), commitThreshold, false, execStage.LogPrefix(), logger),
-				enableChaosMonkey:     execStage.CurrentSyncCycle.IsInitialCycle,
-				hooks:                 hooks,
-				lastCommittedTxNum:    doms.TxNum(),
-				lastCommittedBlockNum: blockNum,
-			},
-		}
-
-		executor = se
 	}
 
-	executor.resetWorkers(ctx, rs, applyTx)
-
 	stepsInDb := rawdbhelpers.IdxStepsCountV3(applyTx)
-	defer func() {
-		executor.LogComplete(stepsInDb)
-	}()
-
-	computeCommitmentDuration := time.Duration(0)
-	blockNum = executor.domains().BlockNum()
+	blockNum = doms.BlockNum()
 
 	if maxBlockNum < blockNum {
 		return nil
@@ -331,255 +264,68 @@ func ExecV3(ctx context.Context,
 	startBlockNum := blockNum
 	blockLimit := uint64(cfg.syncCfg.LoopBlockLimit)
 
-	if blockLimit > 0 && min(blockNum+blockLimit, maxBlockNum) > blockNum+16 || maxBlockNum > blockNum+16 {
-		execType := "serial"
-		if parallel {
-			execType = "parallel"
+	var lastCommittedTxNum uint64
+	var lastCommittedBlockNum uint64
+
+	if parallel {
+		pe := &parallelExecutor{
+			txExecutor: txExecutor{
+				cfg:                   cfg,
+				rs:                    rs,
+				doms:                  doms,
+				agg:                   agg,
+				isMining:              isMining,
+				inMemExec:             inMemExec,
+				logger:                logger,
+				logPrefix:             execStage.LogPrefix(),
+				progress:              NewProgress(blockNum, doms.TxNum(), commitThreshold, false, execStage.LogPrefix(), logger),
+				enableChaosMonkey:     execStage.CurrentSyncCycle.IsInitialCycle,
+				hooks:                 hooks,
+				lastCommittedTxNum:    doms.TxNum(),
+				lastCommittedBlockNum: blockNum,
+			},
+			workerCount: workerCount,
 		}
 
-		log.Info(fmt.Sprintf("[%s] %s starting", execStage.LogPrefix(), execType),
-			"from", blockNum, "to", min(blockNum+blockLimit, maxBlockNum), "fromTxNum", doms.TxNum(), "initialBlockTxOffset", offsetFromBlockBeginning, "initialCycle", initialCycle, "useExternalTx", useExternalTx, "inMem", inMemExec)
-	}
-
-	if !parallel {
-		se := executor.(*serialExecutor)
-
-		execErr = func() error {
-			havePartialBlock := false
-
-			for ; blockNum <= maxBlockNum; blockNum++ {
-				shouldGenerateChangesets := shouldGenerateChangeSets(cfg, blockNum, maxBlockNum, initialCycle)
-				changeSet := &changeset.StateChangeSet{}
-				if shouldGenerateChangesets && blockNum > 0 {
-					executor.domains().SetChangesetAccumulator(changeSet)
-				}
-
-				select {
-				case readAhead <- blockNum:
-				default:
-				}
-
-				b, err = exec3.BlockWithSenders(ctx, cfg.db, applyTx, blockReader, blockNum)
-				if err != nil {
-					return err
-				}
-				if b == nil {
-					// TODO: panic here and see that overall process deadlock
-					return fmt.Errorf("nil block %d", blockNum)
-				}
-
-				txs := b.Transactions()
-				header := b.HeaderNoCopy()
-				totalGasUsed += b.GasUsed()
-				getHashFnMutex := sync.Mutex{}
-
-				blockContext := core.NewEVMBlockContext(header, core.GetHashFn(header, func(hash common.Hash, number uint64) (*types.Header, error) {
-					getHashFnMutex.Lock()
-					defer getHashFnMutex.Unlock()
-					return executor.getHeader(ctx, hash, number)
-				}), cfg.engine, cfg.author, chainConfig)
-
-				if accumulator != nil {
-					txs, err := blockReader.RawTransactions(context.Background(), applyTx, b.NumberU64(), b.NumberU64())
-					if err != nil {
-						return err
-					}
-					accumulator.StartChange(header, txs, false)
-				}
-
-				var txTasks []exec.Task
-
-				for txIndex := -1; txIndex <= len(txs); txIndex++ {
-					// Do not oversend, wait for the result heap to go under certain size
-					txTask := &exec.TxTask{
-						TxNum:           inputTxNum,
-						TxIndex:         txIndex,
-						Header:          header,
-						Uncles:          b.Uncles(),
-						Txs:             txs,
-						EvmBlockContext: blockContext,
-						Withdrawals:     b.Withdrawals(),
-
-						// use history reader instead of state reader to catch up to the tx where we left off
-						HistoryExecution: offsetFromBlockBeginning > 0 && txIndex < int(offsetFromBlockBeginning),
-						Trace:            dbg.TraceTx(blockNum, txIndex),
-						Hooks:            hooks,
-						Logger:           logger,
-					}
-
-					if txTask.TxNum > 0 && txTask.TxNum <= outputTxNum.Load() {
-						havePartialBlock = true
-						inputTxNum++
-						continue
-					}
-
-					txTasks = append(txTasks, txTask)
-					stageProgress = blockNum
-					inputTxNum++
-				}
-
-				continueLoop, err := se.execute(ctx, txTasks, execStage.CurrentSyncCycle.IsInitialCycle, false)
-
-				if err != nil {
-					return err
-				}
-
-				uncommitedGas = uint64(se.executedGas.Load() - int64(se.committedGas))
-
-				if !continueLoop {
-					return nil
-				}
-
-				if !dbg.BatchCommitments || shouldGenerateChangesets {
-					start := time.Now()
-					if dbg.TraceBlock(blockNum) {
-						se.doms.SetTrace(true, false)
-					}
-					rh, err := executor.domains().ComputeCommitment(ctx, applyTx, true, blockNum, inputTxNum, execStage.LogPrefix(), nil)
-					se.doms.SetTrace(false, false)
-
-					if err != nil {
-						return err
-					}
-
-					computeCommitmentDuration += time.Since(start)
-					executor.domains().SavePastChangesetAccumulator(b.Hash(), blockNum, changeSet)
-					if !inMemExec {
-						if err := changeset.WriteDiffSet(applyTx, blockNum, b.Hash(), changeSet); err != nil {
-							return err
-						}
-					}
-					executor.domains().SetChangesetAccumulator(nil)
-
-					if !isMining && !bytes.Equal(rh, header.Root.Bytes()) {
-						logger.Error(fmt.Sprintf("[%s] Wrong trie root of block %d: %x, expected (from header): %x. Block hash: %x", execStage.LogPrefix(), header.Number.Uint64(), rh, header.Root.Bytes(), header.Hash()))
-						return fmt.Errorf("%w, block=%d", ErrWrongTrieRoot, blockNum)
-					}
-				}
-
-				if dbg.StopAfterBlock > 0 && blockNum == dbg.StopAfterBlock {
-					panic(fmt.Sprintf("stopping: block %d complete", blockNum))
-					//return fmt.Errorf("stopping: block %d complete", blockNum)
-				}
-
-				if offsetFromBlockBeginning > 0 {
-					// after history execution no offset will be required
-					offsetFromBlockBeginning = 0
-				}
-
-				select {
-				case <-logEvery.C:
-					if inMemExec || isMining {
-						break
-					}
-
-					executor.LogExecuted()
-
-					//TODO: https://github.com/erigontech/erigon/issues/10724
-					//if executor.tx().(dbstate.HasAggTx).AggTx().(*dbstate.AggregatorRoTx).CanPrune(executor.tx(), outputTxNum.Load()) {
-					//	//small prune cause MDBX_TXN_FULL
-					//	if _, err := executor.tx().(dbstate.HasAggTx).AggTx().(*dbstate.AggregatorRoTx).PruneSmallBatches(ctx, 10*time.Hour, executor.tx()); err != nil {
-					//		return err
-					//	}
-					//}
-
-					isBatchFull := executor.readState().SizeEstimate() >= commitThreshold
-					canPrune := dbstate.AggTx(applyTx).CanPrune(applyTx, outputTxNum.Load())
-					needCalcRoot := isBatchFull || havePartialBlock || canPrune
-					// If we have a partial first block it may not be validated, then we should compute root hash ASAP for fail-fast
-
-					// this will only happen for the first executed block
-					havePartialBlock = false
-
-					if !needCalcRoot {
-						break
-					}
-
-					resetExecGauges(ctx)
-
-					var (
-						commitStart   = time.Now()
-						pruneDuration time.Duration
-					)
-
-					se := executor.(*serialExecutor)
-
-					ok, times, err := flushAndCheckCommitmentV3(ctx, b.HeaderNoCopy(), applyTx, executor.domains(), cfg, execStage, stageProgress, parallel, logger, u, inMemExec)
-					if err != nil {
-						return err
-					} else if !ok {
-						return nil
-					}
-
-					resetCommitmentGauges(ctx)
-
-					computeCommitmentDuration += times.ComputeCommitment
-					flushDuration := times.Flush
-
-					se.txExecutor.lastCommittedBlockNum = b.NumberU64()
-					se.txExecutor.lastCommittedTxNum = inputTxNum
-
-					timeStart := time.Now()
-
-					pruneTimeout := 250 * time.Millisecond
-					if initialCycle {
-						pruneTimeout = 10 * time.Hour
-
-						if err = applyTx.GreedyPruneHistory(ctx, kv.CommitmentDomain); err != nil {
-							return err
-						}
-					}
-
-					if _, err := applyTx.PruneSmallBatches(ctx, pruneTimeout); err != nil {
-						return err
-					}
-
-					pruneDuration = time.Since(timeStart)
-
-					stepsInDb = rawdbhelpers.IdxStepsCountV3(applyTx)
-
-					var commitDuration time.Duration
-					applyTx, commitDuration, err = executor.(*serialExecutor).commit(ctx, execStage, applyTx, nil, useExternalTx)
-					if err != nil {
-						return err
-					}
-					// on chain-tip: if batch is full then stop execution - to allow stages commit
-					if !initialCycle {
-						if isBatchFull {
-							return &ErrLoopExhausted{From: startBlockNum, To: blockNum, Reason: "block batch is full"}
-						}
-
-						if canPrune {
-							return &ErrLoopExhausted{From: startBlockNum, To: blockNum, Reason: "block batch can be pruned"}
-						}
-					}
-
-					if !useExternalTx {
-						executor.LogCommitted(commitStart, 0, 0, uncommitedGas, stepsInDb, commitment.CommitProgress{})
-					}
-
-					uncommitedGas = 0
-
-					logger.Info("Committed", "time", time.Since(commitStart),
-						"block", executor.domains().BlockNum(), "txNum", executor.domains().TxNum(),
-						"step", fmt.Sprintf("%.1f", float64(executor.domains().TxNum())/float64(agg.StepSize())),
-						"flush", flushDuration, "compute commitment", computeCommitmentDuration, "tx.commit", commitDuration, "prune", pruneDuration)
-				default:
-				}
-
-				select {
-				case <-ctx.Done():
-					return ctx.Err()
-				default:
-				}
-
-				if blockLimit > 0 && blockNum-startBlockNum+1 >= blockLimit {
-					return &ErrLoopExhausted{From: startBlockNum, To: blockNum, Reason: "block limit reached"}
-				}
-			}
-
-			return nil
+		defer func() {
+			pe.LogComplete(stepsInDb)
 		}()
+
+		flushEvery := time.NewTicker(2 * time.Second)
+		defer flushEvery.Stop()
+
+		applyTx, execErr = pe.exec(ctx, execStage, u, startBlockNum, offsetFromBlockBeginning, maxBlockNum, blockLimit,
+			inputTxNum, useExternalTx, initialCycle, applyTx, accumulator, readAhead, logEvery, flushEvery)
+
+		lastCommittedBlockNum = pe.lastCommittedBlockNum
+		lastCommittedTxNum = pe.lastCommittedTxNum
+	} else {
+		se := &serialExecutor{
+			txExecutor: txExecutor{
+				cfg:                   cfg,
+				rs:                    rs,
+				doms:                  doms,
+				agg:                   agg,
+				u:                     u,
+				isMining:              isMining,
+				inMemExec:             inMemExec,
+				applyTx:               applyTx,
+				logger:                logger,
+				logPrefix:             execStage.LogPrefix(),
+				progress:              NewProgress(blockNum, doms.TxNum(), commitThreshold, false, execStage.LogPrefix(), logger),
+				enableChaosMonkey:     execStage.CurrentSyncCycle.IsInitialCycle,
+				hooks:                 hooks,
+				lastCommittedTxNum:    doms.TxNum(),
+				lastCommittedBlockNum: blockNum,
+			}}
+
+		defer func() {
+			se.LogComplete(stepsInDb)
+		}()
+
+		b, applyTx, execErr = se.exec(ctx, execStage, u, startBlockNum, offsetFromBlockBeginning, maxBlockNum, blockLimit,
+			inputTxNum, useExternalTx, initialCycle, applyTx, accumulator,
+			readAhead, logEvery)
 
 		if u != nil && !u.HasUnwindPoint() {
 			if b != nil {
@@ -589,14 +335,14 @@ func ExecV3(ctx context.Context,
 							b.NumberU64(), b.Hash(), b.ParentHash(), applyTx, cfg, execStage, maxBlockNum, logger, u)
 					}
 				} else {
-					_, _, err = flushAndCheckCommitmentV3(ctx, b.HeaderNoCopy(), applyTx, executor.domains(), cfg, execStage, stageProgress, parallel, logger, u, inMemExec)
+					_, _, err = flushAndCheckCommitmentV3(ctx, b.HeaderNoCopy(), applyTx, se.domains(), cfg, execStage, stageProgress, parallel, logger, u, inMemExec)
 					if err != nil {
 						return err
 					}
 
-					se.txExecutor.lastCommittedBlockNum = b.NumberU64()
-					committedTransactions := inputTxNum - se.txExecutor.lastCommittedTxNum
-					se.txExecutor.lastCommittedTxNum = inputTxNum
+					se.lastCommittedBlockNum = b.NumberU64()
+					committedTransactions := inputTxNum - se.lastCommittedTxNum
+					se.lastCommittedTxNum = inputTxNum
 
 					commitStart := time.Now()
 					stepsInDb = rawdbhelpers.IdxStepsCountV3(applyTx)
@@ -606,7 +352,7 @@ func ExecV3(ctx context.Context,
 					}
 
 					if !useExternalTx {
-						executor.LogCommitted(commitStart, 0, committedTransactions, uncommitedGas, stepsInDb, commitment.CommitProgress{})
+						se.LogCommitted(commitStart, 0, committedTransactions, uncommitedGas, stepsInDb, commitment.CommitProgress{})
 					}
 
 					uncommitedGas = 0
@@ -615,326 +361,24 @@ func ExecV3(ctx context.Context,
 				fmt.Printf("[dbg] mmmm... do we need action here????\n")
 			}
 		}
-	} else {
-		pe := executor.(*parallelExecutor)
 
-		var asyncTxChan mdbx.TxApplyChan
-		var asyncTx kv.Tx
-
-		switch applyTx := applyTx.(type) {
-		case *temporal.RwTx:
-			temporalTx := applyTx.AsyncClone(mdbx.NewAsyncRwTx(applyTx.RwTx, 1000))
-			asyncTxChan = temporalTx.ApplyChan()
-			asyncTx = temporalTx
-		default:
-			return fmt.Errorf("expected *temporal.RwTx: got %T", applyTx)
-		}
-
-		applyResults := make(chan applyResult, 100_000)
-
-		maxExecBlockNum := maxBlockNum
-		if blockLimit > 0 && blockNum+uint64(blockLimit) < maxBlockNum {
-			maxExecBlockNum = blockNum + blockLimit - 1
-		}
-
-		if err := executor.executeBlocks(executorContext, asyncTx, blockNum, maxExecBlockNum, readAhead, applyResults); err != nil {
-			return err
-		}
-
-		var lastExecutedLog time.Time
-		var lastCommitedLog time.Time
-		var lastBlockResult blockResult
-		var uncommittedBlocks int64
-		var uncommittedTransactions uint64
-		var uncommittedGas int64
-		var flushPending bool
-
-		execErr = func() error {
-			defer func() {
-				if rec := recover(); rec != nil {
-					pe.logger.Warn("["+execStage.LogPrefix()+"] rw panic", "rec", rec, "stack", dbg.Stack())
-				} else if err != nil && !errors.Is(err, context.Canceled) {
-					pe.logger.Warn("["+execStage.LogPrefix()+"] rw exit", "err", err)
-				} else {
-					pe.logger.Debug("[" + execStage.LogPrefix() + "] rw exit")
-				}
-			}()
-
-			shouldGenerateChangesets := shouldGenerateChangeSets(cfg, blockNum, maxBlockNum, initialCycle)
-			changeSet := &changeset.StateChangeSet{}
-			if shouldGenerateChangesets && blockNum > 0 {
-				executor.domains().SetChangesetAccumulator(changeSet)
-			}
-
-			blockUpdateCount := 0
-			blockApplyCount := 0
-
-			for {
-				select {
-				case request := <-asyncTxChan:
-					request.Apply()
-				case applyResult := <-applyResults:
-					switch applyResult := applyResult.(type) {
-					case *txResult:
-						uncommittedGas += applyResult.gasUsed
-						uncommittedTransactions++
-						pe.rs.SetTxNum(applyResult.blockNum, applyResult.txNum)
-						if dbg.TraceApply && dbg.TraceBlock(applyResult.blockNum) {
-							pe.rs.SetTrace(true)
-							fmt.Println(applyResult.blockNum, "apply", applyResult.txNum, applyResult.stateUpdates.UpdateCount())
-						}
-						blockUpdateCount += applyResult.stateUpdates.UpdateCount()
-						err := pe.rs.ApplyTxState(ctx, applyTx, applyResult.blockNum, applyResult.txNum, applyResult.stateUpdates,
-							nil, applyResult.receipt, applyResult.logs, applyResult.traceFroms, applyResult.traceTos,
-							pe.cfg.chainConfig, applyResult.rules, false)
-						blockApplyCount += applyResult.stateUpdates.UpdateCount()
-						pe.rs.SetTrace(false)
-						if err != nil {
-							return err
-						}
-					case *blockResult:
-						if applyResult.BlockNum > 0 && !applyResult.isPartial { //Disable check for genesis. Maybe need somehow improve it in future - to satisfy TestExecutionSpec
-							checkReceipts := !cfg.vmConfig.StatelessExec &&
-								cfg.chainConfig.IsByzantium(applyResult.BlockNum) &&
-								!cfg.vmConfig.NoReceipts && !isMining
-
-							b, err = blockReader.BlockByHash(ctx, applyTx, applyResult.BlockHash)
-
-							if err != nil {
-								return fmt.Errorf("can't retrieve block %d: for post validation: %w", applyResult.BlockNum, err)
-							}
-
-							if b.NumberU64() != applyResult.BlockNum {
-								return fmt.Errorf("block numbers don't match expected: %d: got: %d for hash %x", applyResult.BlockNum, b.NumberU64(), applyResult.BlockHash)
-							}
-
-							if blockUpdateCount != applyResult.ApplyCount {
-								return fmt.Errorf("block %d: applyCount mismatch: got: %d expected %d", applyResult.BlockNum, blockUpdateCount, applyResult.ApplyCount)
-							}
-
-							if err := core.BlockPostValidation(applyResult.GasUsed, applyResult.BlobGasUsed, checkReceipts, applyResult.Receipts,
-								b.HeaderNoCopy(), pe.isMining, b.Transactions(), pe.cfg.chainConfig, pe.logger); err != nil {
-								dumpTxIODebug(applyResult.BlockNum, applyResult.TxIO)
-								return fmt.Errorf("%w, block=%d, %v", consensus.ErrInvalidBlock, applyResult.BlockNum, err) //same as in stage_exec.go
-							}
-
-							if !isMining && !applyResult.isPartial && !execStage.CurrentSyncCycle.IsInitialCycle {
-								cfg.notifications.RecentLogs.Add(applyResult.Receipts)
-							}
-						}
-
-						if applyResult.BlockNum > lastBlockResult.BlockNum {
-							uncommittedBlocks++
-							pe.doms.SetTxNum(applyResult.lastTxNum)
-							pe.doms.SetBlockNum(applyResult.BlockNum)
-							lastBlockResult = *applyResult
-						}
-
-						flushPending = pe.rs.SizeEstimate() > pe.cfg.batchSize.Bytes()
-
-						if !dbg.DiscardCommitment() {
-							if !dbg.BatchCommitments || shouldGenerateChangesets || lastBlockResult.BlockNum == maxExecBlockNum ||
-								(flushPending && lastBlockResult.BlockNum > pe.lastCommittedBlockNum()) {
-
-								resetExecGauges(ctx)
-
-								if dbg.TraceApply && dbg.TraceBlock(applyResult.BlockNum) {
-									fmt.Println(applyResult.BlockNum, "applied count", blockApplyCount, "last tx", applyResult.lastTxNum)
-								}
-
-								var trace bool
-								if dbg.TraceBlock(applyResult.BlockNum) {
-									fmt.Println(applyResult.BlockNum, "Commitment")
-									trace = true
-								}
-								pe.doms.SetTrace(trace, !dbg.BatchCommitments)
-
-								commitProgress := make(chan *commitment.CommitProgress, 100)
-
-								go func() {
-									logEvery := time.NewTicker(20 * time.Second)
-									commitStart := time.Now()
-
-									defer logEvery.Stop()
-									var lastProgress commitment.CommitProgress
-
-									var prevCommitedBlocks uint64
-									var prevCommittedTransactions uint64
-									var prevCommitedGas uint64
-
-									logCommitted := func(commitProgress commitment.CommitProgress) {
-										// this is an approximation of blcok prgress - it assumnes an
-										// even distribution of keys to blocks
-										if commitProgress.KeyIndex > 0 {
-											progress := float64(commitProgress.KeyIndex) / float64(commitProgress.UpdateCount)
-											committedGas := uint64(float64(uncommittedGas) * progress)
-											committedTransactions := uint64(float64(uncommittedTransactions) * progress)
-											commitedBlocks := uint64(float64(uncommittedBlocks) * progress)
-
-											if committedTransactions-prevCommittedTransactions > 0 {
-												pe.LogCommitted(commitStart,
-													commitedBlocks-prevCommitedBlocks,
-													committedTransactions-prevCommittedTransactions,
-													committedGas-prevCommitedGas, stepsInDb, commitProgress)
-											}
-
-											lastCommitedLog = time.Now()
-											prevCommitedBlocks = commitedBlocks
-											prevCommittedTransactions = committedTransactions
-											prevCommitedGas = committedGas
-										}
-
-										if pe.agg.HasBackgroundFilesBuild() {
-											logger.Info(fmt.Sprintf("[%s] Background files build", pe.logPrefix), "progress", pe.agg.BackgroundProgress())
-										}
-									}
-
-									for {
-										select {
-										case <-ctx.Done():
-											return
-										case progress, ok := <-commitProgress:
-											if !ok {
-												if time.Since(lastCommitedLog) > logInterval/20 {
-													logCommitted(lastProgress)
-												}
-												return
-											}
-											lastProgress = *progress
-										case <-logEvery.C:
-											if time.Since(lastCommitedLog) > logInterval-(logInterval/90) {
-												logCommitted(lastProgress)
-											}
-										}
-									}
-								}()
-
-								if time.Since(lastExecutedLog) > logInterval/50 {
-									pe.LogExecuted()
-									lastExecutedLog = time.Now()
-								}
-
-								rh, err := pe.doms.ComputeCommitment(ctx, applyTx, true, applyResult.BlockNum, applyResult.lastTxNum, pe.logPrefix, commitProgress)
-								close(commitProgress)
-								captured := pe.doms.SetTrace(false, false)
-								if err != nil {
-									return err
-								}
-								resetCommitmentGauges(ctx)
-
-								executor.domains().SavePastChangesetAccumulator(applyResult.BlockHash, blockNum, changeSet)
-								if !inMemExec {
-									if err := changeset.WriteDiffSet(applyTx, blockNum, applyResult.BlockHash, changeSet); err != nil {
-										return err
-									}
-								}
-								executor.domains().SetChangesetAccumulator(nil)
-
-								if !bytes.Equal(rh, applyResult.StateRoot.Bytes()) {
-									logger.Error(fmt.Sprintf("[%s] Wrong trie root of block %d: %x, expected (from header): %x. Block hash: %x", pe.logPrefix, applyResult.BlockNum, rh, applyResult.StateRoot.Bytes(), applyResult.BlockHash))
-									if !dbg.BatchCommitments {
-										for _, line := range captured {
-											fmt.Println(line)
-										}
-
-										dumpTxIODebug(applyResult.BlockNum, applyResult.TxIO)
-									}
-
-									return handleIncorrectRootHashError(
-										applyResult.BlockNum, applyResult.BlockHash, applyResult.ParentHash,
-										applyTx, cfg, execStage, maxBlockNum, logger, u)
-								}
-								// fix these here - they will contain estimates after commit logging
-								pe.txExecutor.lastCommittedBlockNum = lastBlockResult.BlockNum
-								pe.txExecutor.lastCommittedTxNum = lastBlockResult.lastTxNum
-								uncommittedBlocks = 0
-								uncommittedGas = 0
-								uncommittedGas = 0
-							}
-						}
-
-						blockUpdateCount = 0
-						blockApplyCount = 0
-
-						if dbg.StopAfterBlock > 0 && applyResult.BlockNum == dbg.StopAfterBlock {
-							return fmt.Errorf("stopping: block %d complete", applyResult.BlockNum)
-						}
-
-						if applyResult.BlockNum == maxExecBlockNum {
-							switch {
-							case applyResult.BlockNum == maxBlockNum:
-								return nil
-							case blockLimit > 0:
-								return &ErrLoopExhausted{From: startBlockNum, To: applyResult.BlockNum, Reason: "block limit reached"}
-							default:
-								return nil
-							}
-						}
-
-						if shouldGenerateChangesets && blockNum > 0 {
-							changeSet = &changeset.StateChangeSet{}
-							executor.domains().SetChangesetAccumulator(changeSet)
-						}
-					}
-				case <-executorContext.Done():
-					err = executor.wait(ctx)
-					return fmt.Errorf("executor context failed: %w", err)
-				case <-ctx.Done():
-					return ctx.Err()
-				case <-logEvery.C:
-					if time.Since(lastExecutedLog) > logInterval-(logInterval/90) {
-						lastExecutedLog = time.Now()
-						pe.LogExecuted()
-						if pe.agg.HasBackgroundFilesBuild() {
-							logger.Info(fmt.Sprintf("[%s] Background files build", pe.logPrefix), "progress", pe.agg.BackgroundProgress())
-						}
-					}
-				case <-flushEvery.C:
-					if flushPending {
-						if !initialCycle {
-							return &ErrLoopExhausted{From: startBlockNum, To: blockNum, Reason: "block batch is full"}
-						}
-
-						if applyTx, err = pe.flushAndCommit(ctx, execStage, applyTx, asyncTxChan, useExternalTx); err != nil {
-							return fmt.Errorf("flush failed: %w", err)
-						}
-
-						flushPending = false
-					}
-				}
-			}
-		}()
-
-		executorCancel()
-
-		if execErr != nil {
-			if !(errors.Is(execErr, context.Canceled) || errors.Is(execErr, &ErrLoopExhausted{})) {
-				return execErr
-			}
-		}
-
-		if applyTx, err = pe.flushAndCommit(ctx, execStage, applyTx, asyncTxChan, useExternalTx); err != nil {
-			return fmt.Errorf("flush failed: %w", err)
-		}
-	}
-
-	if err := executor.wait(ctx); err != nil {
-		return fmt.Errorf("executer wait failed: %w", err)
+		lastCommittedBlockNum = se.lastCommittedBlockNum
+		lastCommittedTxNum = se.lastCommittedTxNum
 	}
 
 	if false && !inMemExec {
-		dumpPlainStateDebug(applyTx, executor.domains())
+		dumpPlainStateDebug(applyTx, doms)
 	}
 
-	lastCommitedStep := kv.Step((executor.lastCommittedTxNum()) / doms.StepSize())
+	lastCommitedStep := kv.Step((lastCommittedTxNum) / doms.StepSize())
 	lastFrozenStep := applyTx.StepsInFiles(kv.CommitmentDomain)
 
 	if lastCommitedStep > 0 && lastCommitedStep <= lastFrozenStep && !dbg.DiscardCommitment() {
 		logger.Warn("["+execStage.LogPrefix()+"] can't persist comittement: txn step frozen",
-			"block", executor.lastCommittedBlockNum(), "txNum", executor.lastCommittedTxNum(), "step", lastCommitedStep,
+			"block", lastCommittedBlockNum, "txNum", lastCommittedTxNum, "step", lastCommitedStep,
 			"lastFrozenStep", lastFrozenStep, "lastFrozenTxNum", ((lastFrozenStep+1)*kv.Step(doms.StepSize()))-1)
 		return fmt.Errorf("can't persist comittement for blockNum %d, txNum %d: step %d is frozen",
-			executor.lastCommittedBlockNum(), executor.lastCommittedTxNum(), lastCommitedStep)
+			lastCommittedBlockNum, lastCommittedTxNum, lastCommitedStep)
 	}
 
 	if !useExternalTx && applyTx != nil {
@@ -943,7 +387,7 @@ func ExecV3(ctx context.Context,
 		}
 	}
 
-	agg.BuildFilesInBackground(outputTxNum.Load())
+	agg.BuildFilesInBackground(doms.TxNum())
 
 	if !shouldReportToTxPool && cfg.notifications != nil && cfg.notifications.Accumulator != nil && !isMining && b != nil {
 		// No reporting to the txn pool has been done since we are not within the "state-stream" window.
@@ -990,6 +434,294 @@ func dumpTxIODebug(blockNum uint64, txIO *state.VersionedIO) {
 			fmt.Println(fmt.Sprintf("%d (%d.%d)", blockNum, txIndex, txIncarnation), "WRT", vw.String())
 		}
 	}
+}
+
+type txExecutor struct {
+	sync.RWMutex
+	cfg              ExecuteBlockCfg
+	agg              *dbstate.Aggregator
+	rs               *state.StateV3Buffered
+	doms             *dbstate.SharedDomains
+	u                Unwinder
+	isMining         bool
+	inMemExec        bool
+	applyTx          kv.TemporalTx
+	logger           log.Logger
+	logPrefix        string
+	progress         *Progress
+	taskExecMetrics  *exec3.WorkerMetrics
+	blockExecMetrics *blockExecMetrics
+	hooks            *tracing.Hooks
+
+	lastExecutedBlockNum  atomic.Int64
+	lastExecutedTxNum     atomic.Int64
+	executedGas           atomic.Int64
+	lastCommittedBlockNum uint64
+	lastCommittedTxNum    uint64
+	committedGas          int64
+
+	execLoopGroup *errgroup.Group
+
+	execRequests chan *execRequest
+	execCount    atomic.Int64
+	abortCount   atomic.Int64
+	invalidCount atomic.Int64
+	readCount    atomic.Int64
+	writeCount   atomic.Int64
+
+	enableChaosMonkey bool
+}
+
+func (te *txExecutor) readState() *state.StateV3Buffered {
+	return te.rs
+}
+
+func (te *txExecutor) domains() *dbstate.SharedDomains {
+	return te.doms
+}
+
+func (te *txExecutor) getHeader(ctx context.Context, hash common.Hash, number uint64) (h *types.Header, err error) {
+	if te.applyTx != nil {
+		err := te.applyTx.Apply(ctx, func(tx kv.Tx) (err error) {
+			h, err = te.cfg.blockReader.Header(ctx, te.applyTx, hash, number)
+			return err
+		})
+
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		if err := te.cfg.db.View(ctx, func(tx kv.Tx) (err error) {
+			h, err = te.cfg.blockReader.Header(ctx, tx, hash, number)
+			return err
+		}); err != nil {
+			return nil, err
+		}
+	}
+
+	return h, nil
+}
+
+func (te *txExecutor) onBlockStart(ctx context.Context, blockNum uint64, blockHash common.Hash) {
+	defer func() {
+		if rec := recover(); rec != nil {
+			te.logger.Warn("hook paniced: %s", rec, "stack", dbg.Stack())
+		}
+	}()
+
+	if te.hooks == nil {
+		return
+	}
+
+	if blockHash == (common.Hash{}) {
+		te.logger.Warn("hooks ignored: zero block hash")
+		return
+	}
+
+	if blockNum == 0 {
+		if te.hooks.OnGenesisBlock != nil {
+			var b *types.Block
+			if err := te.applyTx.Apply(ctx, func(tx kv.Tx) (err error) {
+				b, err = te.cfg.blockReader.BlockByHash(ctx, tx, blockHash)
+				return err
+			}); err != nil {
+				te.logger.Warn("hook: OnGenesisBlock: abandoned", "err", err)
+			}
+			te.hooks.OnGenesisBlock(b, te.cfg.genesis.Alloc)
+		}
+	} else {
+		if te.hooks.OnBlockStart != nil {
+			var b *types.Block
+			var td *big.Int
+			var finalized *types.Header
+			var safe *types.Header
+
+			if err := te.applyTx.Apply(ctx, func(tx kv.Tx) (err error) {
+				b, err = te.cfg.blockReader.BlockByHash(ctx, tx, blockHash)
+				if err != nil {
+					return err
+				}
+				chainReader := NewChainReaderImpl(te.cfg.chainConfig, te.applyTx, te.cfg.blockReader, te.logger)
+				td = chainReader.GetTd(b.ParentHash(), b.NumberU64()-1)
+				finalized = chainReader.CurrentFinalizedHeader()
+				safe = chainReader.CurrentSafeHeader()
+				return nil
+			}); err != nil {
+				te.logger.Warn("hook: OnBlockStart: abandoned", "err", err)
+			}
+
+			te.hooks.OnBlockStart(tracing.BlockEvent{
+				Block:     b,
+				TD:        td,
+				Finalized: finalized,
+				Safe:      safe,
+			})
+		}
+	}
+}
+
+func (te *txExecutor) executeBlocks(ctx context.Context, tx kv.Tx, blockNum uint64, maxBlockNum uint64, readAhead chan uint64, applyResults chan applyResult) error {
+	inputTxNum, _, offsetFromBlockBeginning, err := restoreTxNum(ctx, &te.cfg, tx, te.doms, maxBlockNum)
+
+	if err != nil {
+		return err
+	}
+
+	if te.execLoopGroup == nil {
+		return errors.New("no exec group")
+	}
+
+	te.execLoopGroup.Go(func() (err error) {
+		defer func() {
+			if rec := recover(); rec != nil {
+				err = fmt.Errorf("exec blocks panic: %s", rec)
+			} else if err != nil && !errors.Is(err, context.Canceled) {
+				err = fmt.Errorf("exec blocks error: %w", err)
+			} else {
+				te.logger.Debug("[" + te.logPrefix + "] exec blocks exit")
+			}
+		}()
+
+		for ; blockNum <= maxBlockNum; blockNum++ {
+			select {
+			case readAhead <- blockNum:
+			default:
+			}
+
+			var b *types.Block
+			err := tx.Apply(ctx, func(tx kv.Tx) error {
+				b, err = exec3.BlockWithSenders(ctx, te.cfg.db, tx, te.cfg.blockReader, blockNum)
+				return err
+			})
+			if err != nil {
+				return err
+			}
+			if b == nil {
+				return fmt.Errorf("nil block %d", blockNum)
+			}
+
+			txs := b.Transactions()
+			header := b.HeaderNoCopy()
+			getHashFnMutex := sync.Mutex{}
+
+			blockContext := core.NewEVMBlockContext(header, core.GetHashFn(header, func(hash common.Hash, number uint64) (h *types.Header, err error) {
+				getHashFnMutex.Lock()
+				defer getHashFnMutex.Unlock()
+				err = tx.Apply(ctx, func(tx kv.Tx) (err error) {
+					h, err = te.cfg.blockReader.Header(ctx, tx, hash, number)
+					return err
+				})
+
+				if err != nil {
+					return nil, err
+				}
+
+				return h, err
+			}), te.cfg.engine, te.cfg.author, te.cfg.chainConfig)
+
+			var txTasks []exec.Task
+
+			for txIndex := -1; txIndex <= len(txs); txIndex++ {
+				if inputTxNum > 0 && inputTxNum <= te.progress.initialTxNum {
+					inputTxNum++
+					continue
+				}
+
+				// Do not oversend, wait for the result heap to go under certain size
+				txTask := &exec.TxTask{
+					TxNum:           inputTxNum,
+					TxIndex:         txIndex,
+					Header:          header,
+					Uncles:          b.Uncles(),
+					Txs:             txs,
+					EvmBlockContext: blockContext,
+					Withdrawals:     b.Withdrawals(),
+					// use history reader instead of state reader to catch up to the tx where we left off
+					HistoryExecution: offsetFromBlockBeginning > 0 && txIndex < int(offsetFromBlockBeginning),
+					Config:           te.cfg.chainConfig,
+					Engine:           te.cfg.engine,
+					Trace:            dbg.TraceTx(blockNum, txIndex),
+					Hooks:            te.hooks,
+					Logger:           te.logger,
+				}
+
+				txTasks = append(txTasks, txTask)
+				inputTxNum++
+			}
+
+			te.execRequests <- &execRequest{
+				b.Number().Uint64(), b.Hash(),
+				core.NewGasPool(b.GasLimit(), te.cfg.chainConfig.GetMaxBlobGasPerBlock(b.Time())),
+				txTasks, applyResults, false,
+			}
+
+			mxExecBlocks.Add(1)
+
+			if offsetFromBlockBeginning > 0 {
+				// after history execution no offset will be required
+				offsetFromBlockBeginning = 0
+			}
+		}
+
+		return nil
+	})
+
+	return nil
+}
+
+func (te *txExecutor) commit(ctx context.Context, execStage *StageState, tx kv.TemporalRwTx, useExternalTx bool, resetWorkers func(ctx context.Context, rs *state.StateV3Buffered, applyTx kv.TemporalTx) error) (kv.TemporalRwTx, time.Duration, error) {
+	err := execStage.Update(tx, te.lastCommittedBlockNum)
+
+	if err != nil {
+		return nil, 0, err
+	}
+
+	_, err = rawdb.IncrementStateVersion(tx)
+
+	if err != nil {
+		return nil, 0, fmt.Errorf("writing plain state version: %w", err)
+	}
+
+	tx.CollectMetrics()
+
+	var t2 time.Duration
+
+	if !useExternalTx {
+		tt := time.Now()
+		err = tx.Commit()
+
+		if err != nil {
+			return nil, 0, err
+		}
+
+		t2 = time.Since(tt)
+		dbtx, err := te.cfg.db.BeginRw(ctx) //nolint
+		if err != nil {
+			return nil, t2, err
+		}
+
+		tx = dbtx.(kv.TemporalRwTx)
+	}
+
+	err = resetWorkers(ctx, te.rs, tx)
+
+	if err != nil {
+		if !useExternalTx {
+			tx.Rollback()
+		}
+
+		return nil, t2, err
+	}
+
+	if !useExternalTx {
+		te.agg.BuildFilesInBackground(te.lastCommittedTxNum)
+	}
+
+	if !te.inMemExec {
+		te.doms.ClearRam(false)
+	}
+
+	return tx, t2, nil
 }
 
 // nolint
