+++ execution/stagedsync/exec3_parallel.go
@@ -1,11 +1,11 @@
 package stagedsync
 
 import (
+	"bytes"
 	"context"
 	"errors"
 	"fmt"
 	"maps"
-	"math/big"
 	"sort"
 	"sync"
 	"sync/atomic"
@@ -16,10 +16,14 @@ import (
 	"github.com/erigontech/erigon-lib/metrics"
 	"github.com/erigontech/erigon/db/datadir"
 	"github.com/erigontech/erigon/db/kv/mdbx"
+	"github.com/erigontech/erigon/db/kv/temporal"
+	"github.com/erigontech/erigon/db/rawdb/rawdbhelpers"
+	"github.com/erigontech/erigon/db/state/changeset"
 	"github.com/erigontech/erigon/eth/consensuschain"
 	"github.com/erigontech/erigon/execution/chain"
 	"github.com/erigontech/erigon/execution/commitment"
 	"github.com/erigontech/erigon/execution/exec3/calltracer"
+	"github.com/erigontech/erigon/turbo/shards"
 	"golang.org/x/sync/errgroup"
 
 	"github.com/erigontech/erigon-lib/log/v3"
@@ -29,8 +33,6 @@ import (
 	"github.com/erigontech/erigon/core/tracing"
 	"github.com/erigontech/erigon/core/vm"
 	"github.com/erigontech/erigon/db/kv"
-	"github.com/erigontech/erigon/db/rawdb"
-	dbstate "github.com/erigontech/erigon/db/state"
 	"github.com/erigontech/erigon/execution/consensus"
 	"github.com/erigontech/erigon/execution/exec3"
 	"github.com/erigontech/erigon/execution/tests/chaos_monkey"
@@ -78,25 +80,811 @@ rwloop does:
 When rwLoop has nothing to do - it does Prune, or flush of WAL to RwTx (agg.rotate+agg.Flush)
 */
 
-type executor interface {
-	executeBlocks(ctx context.Context, tx kv.Tx, blockNum uint64, maxBlockNum uint64, readAhead chan uint64, applyResults chan applyResult) error
+type parallelExecutor struct {
+	txExecutor
+	execWorkers    []*exec3.Worker
+	stopWorkers    func()
+	waitWorkers    func()
+	in             *exec.QueueWithRetry
+	rws            *exec.ResultsQueue
+	workerCount    int
+	blockExecutors map[uint64]*blockExecutor
+}
 
-	wait(ctx context.Context) error
-	getHeader(ctx context.Context, hash common.Hash, number uint64) (h *types.Header, err error)
+func (pe *parallelExecutor) exec(ctx context.Context, execStage *StageState, u Unwinder,
+	startBlockNum uint64, offsetFromBlockBeginning uint64, maxBlockNum uint64, blockLimit uint64,
+	inputTxNum uint64, useExternalTx bool, initialCycle bool, rwTx kv.TemporalRwTx, accumulator *shards.Accumulator,
+	readAhead chan uint64, logEvery *time.Ticker, flushEvery *time.Ticker) (kv.TemporalRwTx, error) {
 
-	//these are reset by commit - so need to be read from the executor once its processing
-	readState() *state.StateV3Buffered
-	domains() *dbstate.SharedDomains
+	var asyncTxChan mdbx.TxApplyChan
+	var asyncTx kv.Tx
 
-	commit(ctx context.Context, execStage *StageState, tx kv.TemporalRwTx, asyncTxChan mdbx.TxApplyChan, useExternalTx bool) (kv.TemporalRwTx, time.Duration, error)
-	resetWorkers(ctx context.Context, rs *state.StateV3Buffered, applyTx kv.TemporalTx) error
+	switch applyTx := rwTx.(type) {
+	case *temporal.RwTx:
+		temporalTx := applyTx.AsyncClone(mdbx.NewAsyncRwTx(applyTx.RwTx, 1000))
+		asyncTxChan = temporalTx.ApplyChan()
+		asyncTx = temporalTx
+	default:
+		return rwTx, fmt.Errorf("expected *temporal.RwTx: got %T", rwTx)
+	}
 
-	lastCommittedBlockNum() uint64
-	lastCommittedTxNum() uint64
+	applyResults := make(chan applyResult, 100_000)
 
-	LogExecuted()
-	LogCommitted(commitStart time.Time, committedBlocks uint64, committedTransactions uint64, committedGas uint64, stepsInDb float64, lastProgress commitment.CommitProgress)
-	LogComplete(stepsInDb float64)
+	maxExecBlockNum := maxBlockNum
+	if blockLimit > 0 && startBlockNum+uint64(blockLimit) < maxBlockNum {
+		maxExecBlockNum = startBlockNum + blockLimit - 1
+	}
+
+	if blockLimit > 0 && min(startBlockNum+blockLimit, maxBlockNum) > startBlockNum+16 || maxBlockNum > startBlockNum+16 {
+		log.Info(fmt.Sprintf("[%s] %s starting", execStage.LogPrefix(), "parallel"),
+			"from", startBlockNum, "to", min(startBlockNum+blockLimit, maxBlockNum), "fromTxNum", pe.doms.TxNum(),
+			"initialBlockTxOffset", offsetFromBlockBeginning, "initialCycle", initialCycle, "useExternalTx", useExternalTx,
+			"inMem", pe.inMemExec)
+	}
+
+	executorContext, executorCancel := pe.run(ctx)
+
+	defer executorCancel()
+
+	pe.resetWorkers(ctx, pe.rs, rwTx)
+
+	if err := pe.executeBlocks(executorContext, asyncTx, startBlockNum, maxExecBlockNum, readAhead, applyResults); err != nil {
+		return rwTx, err
+	}
+
+	var lastExecutedLog time.Time
+	var lastCommitedLog time.Time
+	var lastBlockResult blockResult
+	var uncommittedBlocks int64
+	var uncommittedTransactions uint64
+	var uncommittedGas int64
+	var flushPending bool
+
+	execErr := func() (err error) {
+		defer func() {
+			if rec := recover(); rec != nil {
+				pe.logger.Warn("["+execStage.LogPrefix()+"] rw panic", "rec", rec, "stack", dbg.Stack())
+			} else if err != nil && !errors.Is(err, context.Canceled) {
+				pe.logger.Warn("["+execStage.LogPrefix()+"] rw exit", "err", err)
+			} else {
+				pe.logger.Debug("[" + execStage.LogPrefix() + "] rw exit")
+			}
+		}()
+
+		shouldGenerateChangesets := shouldGenerateChangeSets(pe.cfg, startBlockNum, maxBlockNum, initialCycle)
+		changeSet := &changeset.StateChangeSet{}
+		if shouldGenerateChangesets && startBlockNum > 0 {
+			pe.domains().SetChangesetAccumulator(changeSet)
+		}
+
+		blockUpdateCount := 0
+		blockApplyCount := 0
+
+		for {
+			select {
+			case request := <-asyncTxChan:
+				request.Apply()
+			case applyResult := <-applyResults:
+				switch applyResult := applyResult.(type) {
+				case *txResult:
+					uncommittedGas += applyResult.gasUsed
+					uncommittedTransactions++
+					pe.rs.SetTxNum(applyResult.blockNum, applyResult.txNum)
+					if dbg.TraceApply && dbg.TraceBlock(applyResult.blockNum) {
+						pe.rs.SetTrace(true)
+						fmt.Println(applyResult.blockNum, "apply", applyResult.txNum, applyResult.stateUpdates.UpdateCount())
+					}
+					blockUpdateCount += applyResult.stateUpdates.UpdateCount()
+					err := pe.rs.ApplyTxState(ctx, rwTx, applyResult.blockNum, applyResult.txNum, applyResult.stateUpdates,
+						nil, applyResult.receipt, applyResult.logs, applyResult.traceFroms, applyResult.traceTos,
+						pe.cfg.chainConfig, applyResult.rules, false)
+					blockApplyCount += applyResult.stateUpdates.UpdateCount()
+					pe.rs.SetTrace(false)
+					if err != nil {
+						return err
+					}
+				case *blockResult:
+					if applyResult.BlockNum > 0 && !applyResult.isPartial { //Disable check for genesis. Maybe need somehow improve it in future - to satisfy TestExecutionSpec
+						checkReceipts := !pe.cfg.vmConfig.StatelessExec &&
+							pe.cfg.chainConfig.IsByzantium(applyResult.BlockNum) &&
+							!pe.cfg.vmConfig.NoReceipts && !pe.isMining
+
+						b, err := pe.cfg.blockReader.BlockByHash(ctx, rwTx, applyResult.BlockHash)
+
+						if err != nil {
+							return fmt.Errorf("can't retrieve block %d: for post validation: %w", applyResult.BlockNum, err)
+						}
+
+						if b.NumberU64() != applyResult.BlockNum {
+							return fmt.Errorf("block numbers don't match expected: %d: got: %d for hash %x", applyResult.BlockNum, b.NumberU64(), applyResult.BlockHash)
+						}
+
+						if blockUpdateCount != applyResult.ApplyCount {
+							return fmt.Errorf("block %d: applyCount mismatch: got: %d expected %d", applyResult.BlockNum, blockUpdateCount, applyResult.ApplyCount)
+						}
+
+						if err := core.BlockPostValidation(applyResult.GasUsed, applyResult.BlobGasUsed, checkReceipts, applyResult.Receipts,
+							b.HeaderNoCopy(), pe.isMining, b.Transactions(), pe.cfg.chainConfig, pe.logger); err != nil {
+							dumpTxIODebug(applyResult.BlockNum, applyResult.TxIO)
+							return fmt.Errorf("%w, block=%d, %v", consensus.ErrInvalidBlock, applyResult.BlockNum, err) //same as in stage_exec.go
+						}
+
+						if !pe.isMining && !applyResult.isPartial && !execStage.CurrentSyncCycle.IsInitialCycle {
+							pe.cfg.notifications.RecentLogs.Add(applyResult.Receipts)
+						}
+					}
+
+					if applyResult.BlockNum > lastBlockResult.BlockNum {
+						uncommittedBlocks++
+						pe.doms.SetTxNum(applyResult.lastTxNum)
+						pe.doms.SetBlockNum(applyResult.BlockNum)
+						lastBlockResult = *applyResult
+					}
+
+					flushPending = pe.rs.SizeEstimate() > pe.cfg.batchSize.Bytes()
+
+					if !dbg.DiscardCommitment() {
+						if !dbg.BatchCommitments || shouldGenerateChangesets || lastBlockResult.BlockNum == maxExecBlockNum ||
+							(flushPending && lastBlockResult.BlockNum > pe.lastCommittedBlockNum) {
+
+							resetExecGauges(ctx)
+
+							if dbg.TraceApply && dbg.TraceBlock(applyResult.BlockNum) {
+								fmt.Println(applyResult.BlockNum, "applied count", blockApplyCount, "last tx", applyResult.lastTxNum)
+							}
+
+							var trace bool
+							if dbg.TraceBlock(applyResult.BlockNum) {
+								fmt.Println(applyResult.BlockNum, "Commitment")
+								trace = true
+							}
+							pe.doms.SetTrace(trace, !dbg.BatchCommitments)
+
+							commitProgress := make(chan *commitment.CommitProgress, 100)
+
+							go func() {
+								logEvery := time.NewTicker(20 * time.Second)
+								commitStart := time.Now()
+
+								defer logEvery.Stop()
+								var lastProgress commitment.CommitProgress
+
+								var prevCommitedBlocks uint64
+								var prevCommittedTransactions uint64
+								var prevCommitedGas uint64
+
+								logCommitted := func(commitProgress commitment.CommitProgress) {
+									// this is an approximation of blcok prgress - it assumnes an
+									// even distribution of keys to blocks
+									if commitProgress.KeyIndex > 0 {
+										progress := float64(commitProgress.KeyIndex) / float64(commitProgress.UpdateCount)
+										committedGas := uint64(float64(uncommittedGas) * progress)
+										committedTransactions := uint64(float64(uncommittedTransactions) * progress)
+										commitedBlocks := uint64(float64(uncommittedBlocks) * progress)
+
+										if committedTransactions-prevCommittedTransactions > 0 {
+											pe.LogCommitted(commitStart,
+												commitedBlocks-prevCommitedBlocks,
+												committedTransactions-prevCommittedTransactions,
+												committedGas-prevCommitedGas, rawdbhelpers.IdxStepsCountV3(rwTx), commitProgress)
+										}
+
+										lastCommitedLog = time.Now()
+										prevCommitedBlocks = commitedBlocks
+										prevCommittedTransactions = committedTransactions
+										prevCommitedGas = committedGas
+									}
+
+									if pe.agg.HasBackgroundFilesBuild() {
+										pe.logger.Info(fmt.Sprintf("[%s] Background files build", pe.logPrefix), "progress", pe.agg.BackgroundProgress())
+									}
+								}
+
+								for {
+									select {
+									case <-ctx.Done():
+										return
+									case progress, ok := <-commitProgress:
+										if !ok {
+											if time.Since(lastCommitedLog) > logInterval/20 {
+												logCommitted(lastProgress)
+											}
+											return
+										}
+										lastProgress = *progress
+									case <-logEvery.C:
+										if time.Since(lastCommitedLog) > logInterval-(logInterval/90) {
+											logCommitted(lastProgress)
+										}
+									}
+								}
+							}()
+
+							if time.Since(lastExecutedLog) > logInterval/50 {
+								pe.LogExecuted()
+								lastExecutedLog = time.Now()
+							}
+
+							rh, err := pe.doms.ComputeCommitment(ctx, rwTx, true, applyResult.BlockNum, applyResult.lastTxNum, pe.logPrefix, commitProgress)
+							close(commitProgress)
+							captured := pe.doms.SetTrace(false, false)
+							if err != nil {
+								return err
+							}
+							resetCommitmentGauges(ctx)
+
+							pe.domains().SavePastChangesetAccumulator(applyResult.BlockHash, applyResult.BlockNum, changeSet)
+							if !pe.inMemExec {
+								if err := changeset.WriteDiffSet(rwTx, applyResult.BlockNum, applyResult.BlockHash, changeSet); err != nil {
+									return err
+								}
+							}
+							pe.domains().SetChangesetAccumulator(nil)
+
+							if !bytes.Equal(rh, applyResult.StateRoot.Bytes()) {
+								pe.logger.Error(fmt.Sprintf("[%s] Wrong trie root of block %d: %x, expected (from header): %x. Block hash: %x", pe.logPrefix, applyResult.BlockNum, rh, applyResult.StateRoot.Bytes(), applyResult.BlockHash))
+								if !dbg.BatchCommitments {
+									for _, line := range captured {
+										fmt.Println(line)
+									}
+
+									dumpTxIODebug(applyResult.BlockNum, applyResult.TxIO)
+								}
+
+								return handleIncorrectRootHashError(
+									applyResult.BlockNum, applyResult.BlockHash, applyResult.ParentHash,
+									rwTx, pe.cfg, execStage, maxBlockNum, pe.logger, u)
+							}
+							// fix these here - they will contain estimates after commit logging
+							pe.txExecutor.lastCommittedBlockNum = lastBlockResult.BlockNum
+							pe.txExecutor.lastCommittedTxNum = lastBlockResult.lastTxNum
+							uncommittedBlocks = 0
+							uncommittedGas = 0
+							uncommittedGas = 0
+						}
+					}
+
+					blockUpdateCount = 0
+					blockApplyCount = 0
+
+					if dbg.StopAfterBlock > 0 && applyResult.BlockNum == dbg.StopAfterBlock {
+						return fmt.Errorf("stopping: block %d complete", applyResult.BlockNum)
+					}
+
+					if applyResult.BlockNum == maxExecBlockNum {
+						switch {
+						case applyResult.BlockNum == maxBlockNum:
+							return nil
+						case blockLimit > 0:
+							return &ErrLoopExhausted{From: startBlockNum, To: applyResult.BlockNum, Reason: "block limit reached"}
+						default:
+							return nil
+						}
+					}
+
+					if shouldGenerateChangesets && applyResult.BlockNum > 0 {
+						changeSet = &changeset.StateChangeSet{}
+						pe.domains().SetChangesetAccumulator(changeSet)
+					}
+				}
+			case <-executorContext.Done():
+				err = pe.wait(ctx)
+				return fmt.Errorf("executor context failed: %w", err)
+			case <-ctx.Done():
+				return ctx.Err()
+			case <-logEvery.C:
+				if time.Since(lastExecutedLog) > logInterval-(logInterval/90) {
+					lastExecutedLog = time.Now()
+					pe.LogExecuted()
+					if pe.agg.HasBackgroundFilesBuild() {
+						pe.logger.Info(fmt.Sprintf("[%s] Background files build", pe.logPrefix), "progress", pe.agg.BackgroundProgress())
+					}
+				}
+			case <-flushEvery.C:
+				if flushPending {
+					if !initialCycle {
+						return &ErrLoopExhausted{From: startBlockNum, To: lastBlockResult.BlockNum, Reason: "block batch is full"}
+					}
+
+					if rwTx, err = pe.flushAndCommit(ctx, execStage, rwTx, asyncTxChan, useExternalTx); err != nil {
+						return fmt.Errorf("flush failed: %w", err)
+					}
+
+					flushPending = false
+				}
+			}
+		}
+	}()
+
+	executorCancel()
+
+	if execErr != nil {
+		if !(errors.Is(execErr, context.Canceled) || errors.Is(execErr, &ErrLoopExhausted{})) {
+			return rwTx, execErr
+		}
+	}
+
+	var err error
+	if rwTx, err = pe.flushAndCommit(ctx, execStage, rwTx, asyncTxChan, useExternalTx); err != nil {
+		return rwTx, fmt.Errorf("flush failed: %w", err)
+	}
+
+	return rwTx, pe.wait(ctx)
+}
+
+func (pe *parallelExecutor) LogExecuted() {
+	pe.progress.LogExecuted(pe.rs.StateV3, pe)
+	if domainMetrics := pe.domains().LogMetrics(); len(domainMetrics) > 0 {
+		pe.logger.Info(fmt.Sprintf("[%s] domain reads", pe.logPrefix), domainMetrics...)
+	}
+	for domain, domainMetrics := range pe.domains().DomainLogMetrics() {
+		pe.logger.Debug(fmt.Sprintf("[%s] %s", pe.logPrefix, domain), domainMetrics...)
+	}
+}
+
+func (pe *parallelExecutor) LogCommitted(commitStart time.Time, committedBlocks uint64, committedTransactions uint64, committedGas uint64, stepsInDb float64, lastProgress commitment.CommitProgress) {
+	pe.committedGas += int64(committedGas)
+	pe.txExecutor.lastCommittedBlockNum += committedBlocks
+	pe.txExecutor.lastCommittedTxNum += committedTransactions
+	pe.progress.LogCommitted(pe.rs.StateV3, pe, commitStart, stepsInDb, lastProgress)
+	if domainMetrics := pe.domains().LogMetrics(); len(domainMetrics) > 0 {
+		pe.logger.Info(fmt.Sprintf("[%s] domain reads", pe.logPrefix), domainMetrics...)
+	}
+	for domain, domainMetrics := range pe.domains().DomainLogMetrics() {
+		pe.logger.Debug(fmt.Sprintf("[%s] %s", pe.logPrefix, domain), domainMetrics...)
+	}
+}
+
+func (pe *parallelExecutor) LogComplete(stepsInDb float64) {
+	pe.progress.LogComplete(pe.rs.StateV3, pe, stepsInDb)
+	if domainMetrics := pe.domains().LogMetrics(); len(domainMetrics) > 0 {
+		pe.logger.Info(fmt.Sprintf("[%s] domains", pe.logPrefix), domainMetrics...)
+	}
+	for domain, domainMetrics := range pe.domains().DomainLogMetrics() {
+		pe.logger.Debug(fmt.Sprintf("[%s] %s", pe.logPrefix, domain), domainMetrics...)
+	}
+}
+
+func (pe *parallelExecutor) flushAndCommit(ctx context.Context, execStage *StageState, applyTx kv.TemporalRwTx, asyncTxChan mdbx.TxApplyChan, useExternalTx bool) (kv.TemporalRwTx, error) {
+	flushStart := time.Now()
+	var flushTime time.Duration
+
+	if !pe.inMemExec {
+		if err := pe.doms.Flush(ctx, applyTx); err != nil {
+			return applyTx, err
+		}
+		flushTime = time.Since(flushStart)
+	}
+
+	commitStart := time.Now()
+	var t2 time.Duration
+	var err error
+	if applyTx, t2, err = pe.commit(ctx, execStage, applyTx, asyncTxChan, useExternalTx); err != nil {
+		return applyTx, err
+	}
+
+	pe.logger.Info("["+pe.logPrefix+"] flushed", "block", pe.doms.BlockNum(), "time", time.Since(flushStart), "flush", flushTime, "commit", time.Since(commitStart), "db", t2, "externaltx", useExternalTx)
+	return applyTx, nil
+}
+
+func (pe *parallelExecutor) commit(ctx context.Context, execStage *StageState, tx kv.TemporalRwTx, asyncTxChan mdbx.TxApplyChan, useExternalTx bool) (kv.TemporalRwTx, time.Duration, error) {
+	pe.pause()
+	defer pe.resume()
+
+	for {
+		waiter, paused := pe.paused()
+		if paused {
+			break
+		}
+		select {
+		case request := <-asyncTxChan:
+			request.Apply()
+		case <-ctx.Done():
+			return nil, 0, ctx.Err()
+		case <-waiter:
+		}
+	}
+
+	return pe.txExecutor.commit(ctx, execStage, tx, useExternalTx, pe.resetWorkers)
+}
+
+func (pe *parallelExecutor) pause() {
+	for _, worker := range pe.execWorkers {
+		worker.Pause()
+	}
+}
+
+func (pe *parallelExecutor) paused() (chan any, bool) {
+	for _, worker := range pe.execWorkers {
+		if waiter, paused := worker.Paused(); !paused {
+			return waiter, false
+		}
+	}
+
+	return nil, true
+}
+
+func (pe *parallelExecutor) resume() {
+	for _, worker := range pe.execWorkers {
+		worker.Resume()
+	}
+}
+
+func (pe *parallelExecutor) resetWorkers(ctx context.Context, rs *state.StateV3Buffered, _ kv.TemporalTx) error {
+	pe.Lock()
+	defer pe.Unlock()
+
+	for _, worker := range pe.execWorkers {
+		// parallel workers hold thier own tx don't pass in an externals tx
+		worker.ResetState(rs, nil, nil, state.NewNoopWriter(), nil)
+	}
+
+	return nil
+}
+
+func (pe *parallelExecutor) execLoop(ctx context.Context) (err error) {
+	defer func() {
+		pe.Lock()
+		applyTx := pe.applyTx
+		pe.applyTx = nil
+		pe.Unlock()
+
+		if applyTx != nil {
+			applyTx.Rollback()
+		}
+	}()
+
+	defer func() {
+		if rec := recover(); rec != nil {
+			pe.logger.Warn("["+pe.logPrefix+"] exec loop panic", "rec", rec, "stack", dbg.Stack())
+		} else if err != nil && !errors.Is(err, context.Canceled) {
+			pe.logger.Warn("["+pe.logPrefix+"] exec loop error", "err", err)
+		} else {
+			pe.logger.Debug("[" + pe.logPrefix + "] exec loop exit")
+		}
+	}()
+
+	pe.RLock()
+	applyTx := pe.applyTx
+	pe.RUnlock()
+
+	for {
+		err := func() error {
+			pe.Lock()
+			defer pe.Unlock()
+			if applyTx != pe.applyTx {
+				if applyTx != nil {
+					applyTx.Rollback()
+				}
+			}
+
+			if pe.applyTx == nil {
+				temporalDb, ok := pe.cfg.db.(kv.TemporalRwDB)
+				if !ok {
+					return errors.New("pe.cfg.db is not a temporal db")
+				}
+				pe.applyTx, err = temporalDb.BeginTemporalRo(ctx) //nolint
+
+				if err != nil {
+					return err
+				}
+
+				applyTx = pe.applyTx
+			}
+			return nil
+		}()
+
+		if err != nil {
+			return err
+		}
+
+		select {
+		case exec := <-pe.execRequests:
+			if err := pe.processRequest(ctx, exec); err != nil {
+				return err
+			}
+			continue
+		case <-ctx.Done():
+			return ctx.Err()
+		case nextResult, ok := <-pe.rws.ResultCh():
+			if !ok {
+				return nil
+			}
+			closed, err := pe.rws.Drain(ctx, nextResult)
+			if err != nil {
+				return err
+			}
+			if closed {
+				return nil
+			}
+		}
+
+		blockResult, err := pe.processResults(ctx, applyTx)
+
+		if err != nil {
+			return err
+		}
+
+		if blockResult.complete {
+			pe.RLock()
+			blockExecutor, ok := pe.blockExecutors[blockResult.BlockNum]
+			pe.RUnlock()
+
+			if ok {
+				pe.lastExecutedBlockNum.Store(int64(blockResult.BlockNum))
+				pe.execCount.Add(int64(blockExecutor.cntExec))
+				pe.abortCount.Add(int64(blockExecutor.cntAbort))
+				pe.invalidCount.Add(int64(blockExecutor.cntValidationFail))
+				pe.readCount.Add(blockExecutor.blockIO.ReadCount())
+				pe.writeCount.Add(blockExecutor.blockIO.WriteCount())
+
+				blockReceipts := make([]*types.Receipt, 0, len(blockExecutor.results))
+				for _, result := range blockExecutor.results {
+					if result.Receipt != nil {
+						blockReceipts = append(blockReceipts, result.Receipt)
+					}
+				}
+
+				if blockResult.BlockNum > 0 {
+					result := blockExecutor.results[len(blockExecutor.results)-1]
+
+					stateUpdates, err := func() (state.StateUpdates, error) {
+						pe.RLock()
+						defer pe.RUnlock()
+
+						ibs := state.New(state.NewBufferedReader(pe.rs, state.NewReaderV3(pe.rs.Domains().AsGetter(applyTx))))
+						ibs.SetTxContext(result.Version().BlockNum, result.Version().TxIndex)
+						ibs.SetVersion(result.Version().Incarnation)
+
+						txTask, ok := result.Task.(*taskVersion).Task.(*exec.TxTask)
+
+						if !ok {
+							return state.StateUpdates{}, nil
+						}
+
+						syscall := func(contract common.Address, data []byte) ([]byte, error) {
+							ret, err := core.SysCallContract(contract, data, pe.cfg.chainConfig, ibs, txTask.Header, pe.cfg.engine, false, *pe.cfg.vmConfig)
+							if err != nil {
+								return nil, err
+							}
+							result.Logs = append(result.Logs, ibs.GetRawLogs(txTask.TxIndex)...)
+							return ret, err
+						}
+
+						chainReader := consensuschain.NewReader(pe.cfg.chainConfig, applyTx, pe.cfg.blockReader, pe.logger)
+						if pe.isMining {
+							_, _, err =
+								pe.cfg.engine.FinalizeAndAssemble(
+									pe.cfg.chainConfig, types.CopyHeader(txTask.Header), ibs, txTask.Txs, txTask.Uncles, blockReceipts,
+									txTask.Withdrawals, chainReader, syscall, nil, pe.logger)
+						} else {
+							_, err =
+								pe.cfg.engine.Finalize(
+									pe.cfg.chainConfig, types.CopyHeader(txTask.Header), ibs, txTask.Txs, txTask.Uncles, blockReceipts,
+									txTask.Withdrawals, chainReader, syscall, false, pe.logger)
+						}
+
+						if err != nil {
+							return state.StateUpdates{}, fmt.Errorf("can't finalize block: %w", err)
+						}
+
+						stateWriter := state.NewBufferedWriter(pe.rs, nil)
+
+						if err = ibs.MakeWriteSet(txTask.EvmBlockContext.Rules(txTask.Config), stateWriter); err != nil {
+							return state.StateUpdates{}, err
+						}
+
+						return stateWriter.WriteSet(), nil
+					}()
+
+					if err != nil {
+						return err
+					}
+
+					blockResult.ApplyCount += stateUpdates.UpdateCount()
+					if dbg.TraceApply && dbg.TraceBlock(blockResult.BlockNum) {
+						stateUpdates.TraceBlockUpdates(blockResult.BlockNum, true)
+						fmt.Println(blockResult.BlockNum, "apply count", blockResult.ApplyCount)
+					}
+
+					blockExecutor.applyResults <- &txResult{
+						blockNum:     blockResult.BlockNum,
+						txNum:        blockResult.lastTxNum,
+						rules:        result.Rules(),
+						stateUpdates: stateUpdates,
+						logs:         result.Logs,
+						traceFroms:   result.TraceFroms,
+						traceTos:     result.TraceTos,
+					}
+				}
+
+				if !blockExecutor.execStarted.IsZero() {
+					pe.blockExecMetrics.Duration.Add(time.Since(blockExecutor.execStarted))
+					pe.blockExecMetrics.BlockCount.Add(1)
+				}
+				blockExecutor.applyResults <- blockResult
+				pe.Lock()
+				delete(pe.blockExecutors, blockResult.BlockNum)
+				pe.Unlock()
+			}
+
+			pe.RLock()
+			blockExecutor, ok = pe.blockExecutors[blockResult.BlockNum+1]
+			pe.RUnlock()
+
+			if ok {
+				pe.onBlockStart(ctx, blockExecutor.blockNum, blockExecutor.blockHash)
+				blockExecutor.execStarted = time.Now()
+				blockExecutor.scheduleExecution(ctx, pe)
+			}
+		}
+	}
+}
+
+func (pe *parallelExecutor) processRequest(ctx context.Context, execRequest *execRequest) (err error) {
+	prevSenderTx := map[common.Address]int{}
+	var scheduleable *blockExecutor
+	var executor *blockExecutor
+
+	for i, txTask := range execRequest.tasks {
+		t := &execTask{
+			Task:               txTask,
+			index:              i,
+			shouldDelayFeeCalc: true,
+		}
+
+		blockNum := t.Version().BlockNum
+
+		if executor == nil {
+			var ok bool
+			executor, ok = pe.blockExecutors[blockNum]
+
+			if !ok {
+				executor = newBlockExec(blockNum, execRequest.blockHash, execRequest.gasPool, execRequest.applyResults, execRequest.profile)
+			}
+		}
+
+		executor.tasks = append(executor.tasks, t)
+		executor.results = append(executor.results, nil)
+		executor.txIncarnations = append(executor.txIncarnations, 0)
+		executor.execFailed = append(executor.execFailed, 0)
+		executor.execAborted = append(executor.execAborted, 0)
+
+		executor.skipCheck[len(executor.tasks)-1] = false
+		executor.estimateDeps[len(executor.tasks)-1] = []int{}
+
+		executor.execTasks.pushPending(i)
+		executor.validateTasks.pushPending(i)
+
+		if len(t.Dependencies()) > 0 {
+			for _, depTxIndex := range t.Dependencies() {
+				executor.execTasks.addDependency(depTxIndex+1, i)
+			}
+			executor.execTasks.clearPending(i)
+		} else {
+			sender, err := t.TxSender()
+			if err != nil {
+				return err
+			}
+			if sender != nil {
+				if tx, ok := prevSenderTx[*sender]; ok {
+					executor.execTasks.addDependency(tx, i)
+					executor.execTasks.clearPending(i)
+				}
+
+				prevSenderTx[*sender] = i
+			}
+		}
+
+		if t.IsBlockEnd() {
+			pe.Lock()
+			if len(pe.blockExecutors) == 0 {
+				pe.blockExecutors = map[uint64]*blockExecutor{
+					blockNum: executor,
+				}
+				scheduleable = executor
+			} else {
+				pe.blockExecutors[t.Version().BlockNum] = executor
+			}
+			pe.Unlock()
+
+			executor = nil
+		}
+	}
+
+	if scheduleable != nil {
+		pe.blockExecMetrics.BlockCount.Add(1)
+		scheduleable.execStarted = time.Now()
+		scheduleable.scheduleExecution(ctx, pe)
+	}
+
+	return nil
+}
+
+func (pe *parallelExecutor) processResults(ctx context.Context, applyTx kv.TemporalTx) (blockResult *blockResult, err error) {
+	rwsIt := pe.rws.Iter()
+	for rwsIt.HasNext() && (blockResult == nil || !blockResult.complete) {
+		txResult := rwsIt.PopNext()
+
+		if pe.cfg.syncCfg.ChaosMonkey && pe.enableChaosMonkey {
+			chaosErr := chaos_monkey.ThrowRandomConsensusError(false, txResult.Version().TxIndex, pe.cfg.badBlockHalt, txResult.Err)
+			if chaosErr != nil {
+				log.Warn("Monkey in consensus")
+				return blockResult, chaosErr
+			}
+		}
+
+		pe.RLock()
+		blockExecutor, ok := pe.blockExecutors[txResult.Version().BlockNum]
+		pe.RUnlock()
+
+		if !ok {
+			return nil, fmt.Errorf("unknown block: %d", txResult.Version().BlockNum)
+		}
+
+		blockResult, err = blockExecutor.nextResult(ctx, pe, txResult, applyTx)
+
+		if err != nil {
+			return blockResult, err
+		}
+	}
+
+	return blockResult, nil
+}
+
+func (pe *parallelExecutor) run(ctx context.Context) (context.Context, context.CancelFunc) {
+	pe.execRequests = make(chan *execRequest, 100_000)
+	pe.in = exec.NewQueueWithRetry(100_000)
+
+	pe.taskExecMetrics = exec3.NewWorkerMetrics()
+	pe.blockExecMetrics = newBlockExecMetrics()
+
+	execLoopCtx, execLoopCtxCancel := context.WithCancel(ctx)
+	pe.execLoopGroup, execLoopCtx = errgroup.WithContext(execLoopCtx)
+
+	pe.execWorkers, _, pe.rws, pe.stopWorkers, pe.waitWorkers = exec3.NewWorkersPool(
+		execLoopCtx, nil, true, pe.cfg.db, nil, nil, nil, pe.in,
+		pe.cfg.blockReader, pe.cfg.chainConfig, pe.cfg.genesis, pe.cfg.engine,
+		pe.workerCount+1, pe.taskExecMetrics, pe.cfg.dirs, pe.isMining, pe.logger)
+
+	pe.execLoopGroup.Go(func() error {
+		defer pe.rws.Close()
+		defer pe.in.Close()
+		pe.resetWorkers(execLoopCtx, pe.rs, nil)
+		return pe.execLoop(execLoopCtx)
+	})
+
+	return execLoopCtx, func() {
+		execLoopCtxCancel()
+		pe.wait(ctx)
+		pe.stopWorkers()
+		pe.in.Close()
+	}
+}
+
+func (pe *parallelExecutor) wait(ctx context.Context) error {
+	doneCh := make(chan error, 1)
+
+	go func() {
+		if pe.execLoopGroup != nil {
+			err := pe.execLoopGroup.Wait()
+			if err != nil && !errors.Is(err, context.Canceled) {
+				doneCh <- err
+				return
+			}
+			pe.waitWorkers()
+		}
+		doneCh <- nil
+	}()
+
+	for {
+		select {
+		case <-ctx.Done():
+			return nil
+		case err := <-doneCh:
+			return err
+		}
+	}
 }
 
 type applyResult interface {
@@ -352,294 +1140,6 @@ type blockDuration struct {
 func (d *blockDuration) Add(i time.Duration) {
 	d.Int64.Add(int64(i))
 	d.Ema.Update(i)
-}
-
-type txExecutor struct {
-	sync.RWMutex
-	cfg              ExecuteBlockCfg
-	agg              *dbstate.Aggregator
-	rs               *state.StateV3Buffered
-	doms             *dbstate.SharedDomains
-	u                Unwinder
-	isMining         bool
-	inMemExec        bool
-	applyTx          kv.TemporalTx
-	logger           log.Logger
-	logPrefix        string
-	progress         *Progress
-	taskExecMetrics  *exec3.WorkerMetrics
-	blockExecMetrics *blockExecMetrics
-	hooks            *tracing.Hooks
-
-	lastExecutedBlockNum  atomic.Int64
-	lastExecutedTxNum     atomic.Int64
-	executedGas           atomic.Int64
-	lastCommittedBlockNum uint64
-	lastCommittedTxNum    uint64
-	committedGas          int64
-
-	execLoopGroup *errgroup.Group
-
-	execRequests chan *execRequest
-	execCount    atomic.Int64
-	abortCount   atomic.Int64
-	invalidCount atomic.Int64
-	readCount    atomic.Int64
-	writeCount   atomic.Int64
-
-	enableChaosMonkey bool
-}
-
-func (te *txExecutor) readState() *state.StateV3Buffered {
-	return te.rs
-}
-
-func (te *txExecutor) domains() *dbstate.SharedDomains {
-	return te.doms
-}
-
-func (te *txExecutor) getHeader(ctx context.Context, hash common.Hash, number uint64) (h *types.Header, err error) {
-	if te.applyTx != nil {
-		err := te.applyTx.Apply(ctx, func(tx kv.Tx) (err error) {
-			h, err = te.cfg.blockReader.Header(ctx, te.applyTx, hash, number)
-			return err
-		})
-
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		if err := te.cfg.db.View(ctx, func(tx kv.Tx) (err error) {
-			h, err = te.cfg.blockReader.Header(ctx, tx, hash, number)
-			return err
-		}); err != nil {
-			return nil, err
-		}
-	}
-
-	return h, nil
-}
-
-func (te *txExecutor) onBlockStart(ctx context.Context, blockNum uint64, blockHash common.Hash) {
-	defer func() {
-		if rec := recover(); rec != nil {
-			te.logger.Warn("hook paniced: %s", rec, "stack", dbg.Stack())
-		}
-	}()
-
-	if te.hooks == nil {
-		return
-	}
-
-	if blockHash == (common.Hash{}) {
-		te.logger.Warn("hooks ignored: zero block hash")
-		return
-	}
-
-	if blockNum == 0 {
-		if te.hooks.OnGenesisBlock != nil {
-			var b *types.Block
-			if err := te.applyTx.Apply(ctx, func(tx kv.Tx) (err error) {
-				b, err = te.cfg.blockReader.BlockByHash(ctx, tx, blockHash)
-				return err
-			}); err != nil {
-				te.logger.Warn("hook: OnGenesisBlock: abandoned", "err", err)
-			}
-			te.hooks.OnGenesisBlock(b, te.cfg.genesis.Alloc)
-		}
-	} else {
-		if te.hooks.OnBlockStart != nil {
-			var b *types.Block
-			var td *big.Int
-			var finalized *types.Header
-			var safe *types.Header
-
-			if err := te.applyTx.Apply(ctx, func(tx kv.Tx) (err error) {
-				b, err = te.cfg.blockReader.BlockByHash(ctx, tx, blockHash)
-				if err != nil {
-					return err
-				}
-				chainReader := NewChainReaderImpl(te.cfg.chainConfig, te.applyTx, te.cfg.blockReader, te.logger)
-				td = chainReader.GetTd(b.ParentHash(), b.NumberU64()-1)
-				finalized = chainReader.CurrentFinalizedHeader()
-				safe = chainReader.CurrentSafeHeader()
-				return nil
-			}); err != nil {
-				te.logger.Warn("hook: OnBlockStart: abandoned", "err", err)
-			}
-
-			te.hooks.OnBlockStart(tracing.BlockEvent{
-				Block:     b,
-				TD:        td,
-				Finalized: finalized,
-				Safe:      safe,
-			})
-		}
-	}
-}
-
-func (te *txExecutor) executeBlocks(ctx context.Context, tx kv.Tx, blockNum uint64, maxBlockNum uint64, readAhead chan uint64, applyResults chan applyResult) error {
-	inputTxNum, _, offsetFromBlockBeginning, err := restoreTxNum(ctx, &te.cfg, tx, te.doms, maxBlockNum)
-
-	if err != nil {
-		return err
-	}
-
-	if te.execLoopGroup == nil {
-		return errors.New("no exec group")
-	}
-
-	te.execLoopGroup.Go(func() (err error) {
-		defer func() {
-			if rec := recover(); rec != nil {
-				err = fmt.Errorf("exec blocks panic: %s", rec)
-			} else if err != nil && !errors.Is(err, context.Canceled) {
-				err = fmt.Errorf("exec blocks error: %w", err)
-			} else {
-				te.logger.Debug("[" + te.logPrefix + "] exec blocks exit")
-			}
-		}()
-
-		for ; blockNum <= maxBlockNum; blockNum++ {
-			select {
-			case readAhead <- blockNum:
-			default:
-			}
-
-			var b *types.Block
-			err := tx.Apply(ctx, func(tx kv.Tx) error {
-				b, err = exec3.BlockWithSenders(ctx, te.cfg.db, tx, te.cfg.blockReader, blockNum)
-				return err
-			})
-			if err != nil {
-				return err
-			}
-			if b == nil {
-				return fmt.Errorf("nil block %d", blockNum)
-			}
-
-			txs := b.Transactions()
-			header := b.HeaderNoCopy()
-			getHashFnMutex := sync.Mutex{}
-
-			blockContext := core.NewEVMBlockContext(header, core.GetHashFn(header, func(hash common.Hash, number uint64) (h *types.Header, err error) {
-				getHashFnMutex.Lock()
-				defer getHashFnMutex.Unlock()
-				err = tx.Apply(ctx, func(tx kv.Tx) (err error) {
-					h, err = te.cfg.blockReader.Header(ctx, tx, hash, number)
-					return err
-				})
-
-				if err != nil {
-					return nil, err
-				}
-
-				return h, err
-			}), te.cfg.engine, te.cfg.author, te.cfg.chainConfig)
-
-			var txTasks []exec.Task
-
-			for txIndex := -1; txIndex <= len(txs); txIndex++ {
-				if inputTxNum > 0 && inputTxNum <= te.progress.initialTxNum {
-					inputTxNum++
-					continue
-				}
-
-				// Do not oversend, wait for the result heap to go under certain size
-				txTask := &exec.TxTask{
-					TxNum:           inputTxNum,
-					TxIndex:         txIndex,
-					Header:          header,
-					Uncles:          b.Uncles(),
-					Txs:             txs,
-					EvmBlockContext: blockContext,
-					Withdrawals:     b.Withdrawals(),
-					// use history reader instead of state reader to catch up to the tx where we left off
-					HistoryExecution: offsetFromBlockBeginning > 0 && txIndex < int(offsetFromBlockBeginning),
-					Config:           te.cfg.chainConfig,
-					Engine:           te.cfg.engine,
-					Trace:            dbg.TraceTx(blockNum, txIndex),
-					Hooks:            te.hooks,
-					Logger:           te.logger,
-				}
-
-				txTasks = append(txTasks, txTask)
-				inputTxNum++
-			}
-
-			te.execRequests <- &execRequest{
-				b.Number().Uint64(), b.Hash(),
-				core.NewGasPool(b.GasLimit(), te.cfg.chainConfig.GetMaxBlobGasPerBlock(b.Time())),
-				txTasks, applyResults, false,
-			}
-
-			mxExecBlocks.Add(1)
-
-			if offsetFromBlockBeginning > 0 {
-				// after history execution no offset will be required
-				offsetFromBlockBeginning = 0
-			}
-		}
-
-		return nil
-	})
-
-	return nil
-}
-
-func (te *txExecutor) commit(ctx context.Context, execStage *StageState, tx kv.TemporalRwTx, useExternalTx bool, resetWorkers func(ctx context.Context, rs *state.StateV3Buffered, applyTx kv.TemporalTx) error) (kv.TemporalRwTx, time.Duration, error) {
-	err := execStage.Update(tx, te.lastCommittedBlockNum)
-
-	if err != nil {
-		return nil, 0, err
-	}
-
-	_, err = rawdb.IncrementStateVersion(tx)
-
-	if err != nil {
-		return nil, 0, fmt.Errorf("writing plain state version: %w", err)
-	}
-
-	tx.CollectMetrics()
-
-	var t2 time.Duration
-
-	if !useExternalTx {
-		tt := time.Now()
-		err = tx.Commit()
-
-		if err != nil {
-			return nil, 0, err
-		}
-
-		t2 = time.Since(tt)
-		dbtx, err := te.cfg.db.BeginRw(ctx) //nolint
-		if err != nil {
-			return nil, t2, err
-		}
-
-		tx = dbtx.(kv.TemporalRwTx)
-	}
-
-	err = resetWorkers(ctx, te.rs, tx)
-
-	if err != nil {
-		if !useExternalTx {
-			tx.Rollback()
-		}
-
-		return nil, t2, err
-	}
-
-	if !useExternalTx {
-		te.agg.BuildFilesInBackground(te.lastCommittedTxNum)
-	}
-
-	if !te.inMemExec {
-		te.doms.ClearRam(false)
-	}
-
-	return tx, t2, nil
 }
 
 type execRequest struct {
@@ -1150,500 +1650,6 @@ func (be *blockExecutor) scheduleExecution(ctx context.Context, pe *parallelExec
 				profile:    be.profile,
 				stats:      be.stats,
 				statsMutex: &be.Mutex})
-		}
-	}
-}
-
-type parallelExecutor struct {
-	txExecutor
-	execWorkers    []*exec3.Worker
-	stopWorkers    func()
-	waitWorkers    func()
-	in             *exec.QueueWithRetry
-	rws            *exec.ResultsQueue
-	workerCount    int
-	blockExecutors map[uint64]*blockExecutor
-}
-
-func (pe *parallelExecutor) LogExecuted() {
-	pe.progress.LogExecuted(pe.rs.StateV3, pe)
-	if domainMetrics := pe.domains().LogMetrics(); len(domainMetrics) > 0 {
-		pe.logger.Info(fmt.Sprintf("[%s] domain reads", pe.logPrefix), domainMetrics...)
-	}
-	for domain, domainMetrics := range pe.domains().DomainLogMetrics() {
-		pe.logger.Debug(fmt.Sprintf("[%s] %s", pe.logPrefix, domain), domainMetrics...)
-	}
-}
-
-func (pe *parallelExecutor) LogCommitted(commitStart time.Time, committedBlocks uint64, committedTransactions uint64, committedGas uint64, stepsInDb float64, lastProgress commitment.CommitProgress) {
-	pe.committedGas += int64(committedGas)
-	pe.txExecutor.lastCommittedBlockNum += committedBlocks
-	pe.txExecutor.lastCommittedTxNum += committedTransactions
-	pe.progress.LogCommitted(pe.rs.StateV3, pe, commitStart, stepsInDb, lastProgress)
-	if domainMetrics := pe.domains().LogMetrics(); len(domainMetrics) > 0 {
-		pe.logger.Info(fmt.Sprintf("[%s] domain reads", pe.logPrefix), domainMetrics...)
-	}
-	for domain, domainMetrics := range pe.domains().DomainLogMetrics() {
-		pe.logger.Debug(fmt.Sprintf("[%s] %s", pe.logPrefix, domain), domainMetrics...)
-	}
-}
-
-func (pe *parallelExecutor) LogComplete(stepsInDb float64) {
-	pe.progress.LogComplete(pe.rs.StateV3, pe, stepsInDb)
-	if domainMetrics := pe.domains().LogMetrics(); len(domainMetrics) > 0 {
-		pe.logger.Info(fmt.Sprintf("[%s] domains", pe.logPrefix), domainMetrics...)
-	}
-	for domain, domainMetrics := range pe.domains().DomainLogMetrics() {
-		pe.logger.Debug(fmt.Sprintf("[%s] %s", pe.logPrefix, domain), domainMetrics...)
-	}
-}
-
-func (pe *parallelExecutor) lastCommittedBlockNum() uint64 {
-	return pe.txExecutor.lastCommittedBlockNum
-}
-
-func (pe *parallelExecutor) lastCommittedTxNum() uint64 {
-	return pe.txExecutor.lastCommittedTxNum
-}
-
-func (pe *parallelExecutor) flushAndCommit(ctx context.Context, execStage *StageState, applyTx kv.TemporalRwTx, asyncTxChan mdbx.TxApplyChan, useExternalTx bool) (kv.TemporalRwTx, error) {
-	flushStart := time.Now()
-	var flushTime time.Duration
-
-	if !pe.inMemExec {
-		if err := pe.doms.Flush(ctx, applyTx); err != nil {
-			return applyTx, err
-		}
-		flushTime = time.Since(flushStart)
-	}
-
-	commitStart := time.Now()
-	var t2 time.Duration
-	var err error
-	if applyTx, t2, err = pe.commit(ctx, execStage, applyTx, asyncTxChan, useExternalTx); err != nil {
-		return applyTx, err
-	}
-
-	pe.logger.Info("["+pe.logPrefix+"] flushed", "block", pe.doms.BlockNum(), "time", time.Since(flushStart), "flush", flushTime, "commit", time.Since(commitStart), "db", t2, "externaltx", useExternalTx)
-	return applyTx, nil
-}
-
-func (pe *parallelExecutor) commit(ctx context.Context, execStage *StageState, tx kv.TemporalRwTx, asyncTxChan mdbx.TxApplyChan, useExternalTx bool) (kv.TemporalRwTx, time.Duration, error) {
-	pe.pause()
-	defer pe.resume()
-
-	for {
-		waiter, paused := pe.paused()
-		if paused {
-			break
-		}
-		select {
-		case request := <-asyncTxChan:
-			request.Apply()
-		case <-ctx.Done():
-			return nil, 0, ctx.Err()
-		case <-waiter:
-		}
-	}
-
-	return pe.txExecutor.commit(ctx, execStage, tx, useExternalTx, pe.resetWorkers)
-}
-
-func (pe *parallelExecutor) pause() {
-	for _, worker := range pe.execWorkers {
-		worker.Pause()
-	}
-}
-
-func (pe *parallelExecutor) paused() (chan any, bool) {
-	for _, worker := range pe.execWorkers {
-		if waiter, paused := worker.Paused(); !paused {
-			return waiter, false
-		}
-	}
-
-	return nil, true
-}
-
-func (pe *parallelExecutor) resume() {
-	for _, worker := range pe.execWorkers {
-		worker.Resume()
-	}
-}
-
-func (pe *parallelExecutor) resetWorkers(ctx context.Context, rs *state.StateV3Buffered, _ kv.TemporalTx) error {
-	pe.Lock()
-	defer pe.Unlock()
-
-	for _, worker := range pe.execWorkers {
-		// parallel workers hold thier own tx don't pass in an externals tx
-		worker.ResetState(rs, nil, nil, state.NewNoopWriter(), nil)
-	}
-
-	return nil
-}
-
-func (pe *parallelExecutor) execLoop(ctx context.Context) (err error) {
-	defer func() {
-		pe.Lock()
-		applyTx := pe.applyTx
-		pe.applyTx = nil
-		pe.Unlock()
-
-		if applyTx != nil {
-			applyTx.Rollback()
-		}
-	}()
-
-	defer func() {
-		if rec := recover(); rec != nil {
-			pe.logger.Warn("["+pe.logPrefix+"] exec loop panic", "rec", rec, "stack", dbg.Stack())
-		} else if err != nil && !errors.Is(err, context.Canceled) {
-			pe.logger.Warn("["+pe.logPrefix+"] exec loop error", "err", err)
-		} else {
-			pe.logger.Debug("[" + pe.logPrefix + "] exec loop exit")
-		}
-	}()
-
-	pe.RLock()
-	applyTx := pe.applyTx
-	pe.RUnlock()
-
-	for {
-		err := func() error {
-			pe.Lock()
-			defer pe.Unlock()
-			if applyTx != pe.applyTx {
-				if applyTx != nil {
-					applyTx.Rollback()
-				}
-			}
-
-			if pe.applyTx == nil {
-				temporalDb, ok := pe.cfg.db.(kv.TemporalRwDB)
-				if !ok {
-					return errors.New("pe.cfg.db is not a temporal db")
-				}
-				pe.applyTx, err = temporalDb.BeginTemporalRo(ctx) //nolint
-
-				if err != nil {
-					return err
-				}
-
-				applyTx = pe.applyTx
-			}
-			return nil
-		}()
-
-		if err != nil {
-			return err
-		}
-
-		select {
-		case exec := <-pe.execRequests:
-			if err := pe.processRequest(ctx, exec); err != nil {
-				return err
-			}
-			continue
-		case <-ctx.Done():
-			return ctx.Err()
-		case nextResult, ok := <-pe.rws.ResultCh():
-			if !ok {
-				return nil
-			}
-			closed, err := pe.rws.Drain(ctx, nextResult)
-			if err != nil {
-				return err
-			}
-			if closed {
-				return nil
-			}
-		}
-
-		blockResult, err := pe.processResults(ctx, applyTx)
-
-		if err != nil {
-			return err
-		}
-
-		if blockResult.complete {
-			pe.RLock()
-			blockExecutor, ok := pe.blockExecutors[blockResult.BlockNum]
-			pe.RUnlock()
-
-			if ok {
-				pe.lastExecutedBlockNum.Store(int64(blockResult.BlockNum))
-				pe.execCount.Add(int64(blockExecutor.cntExec))
-				pe.abortCount.Add(int64(blockExecutor.cntAbort))
-				pe.invalidCount.Add(int64(blockExecutor.cntValidationFail))
-				pe.readCount.Add(blockExecutor.blockIO.ReadCount())
-				pe.writeCount.Add(blockExecutor.blockIO.WriteCount())
-
-				blockReceipts := make([]*types.Receipt, 0, len(blockExecutor.results))
-				for _, result := range blockExecutor.results {
-					if result.Receipt != nil {
-						blockReceipts = append(blockReceipts, result.Receipt)
-					}
-				}
-
-				if blockResult.BlockNum > 0 {
-					result := blockExecutor.results[len(blockExecutor.results)-1]
-
-					stateUpdates, err := func() (state.StateUpdates, error) {
-						pe.RLock()
-						defer pe.RUnlock()
-
-						ibs := state.New(state.NewBufferedReader(pe.rs, state.NewReaderV3(pe.rs.Domains().AsGetter(applyTx))))
-						ibs.SetTxContext(result.Version().BlockNum, result.Version().TxIndex)
-						ibs.SetVersion(result.Version().Incarnation)
-
-						txTask, ok := result.Task.(*taskVersion).Task.(*exec.TxTask)
-
-						if !ok {
-							return state.StateUpdates{}, nil
-						}
-
-						syscall := func(contract common.Address, data []byte) ([]byte, error) {
-							ret, err := core.SysCallContract(contract, data, pe.cfg.chainConfig, ibs, txTask.Header, pe.cfg.engine, false, *pe.cfg.vmConfig)
-							if err != nil {
-								return nil, err
-							}
-							result.Logs = append(result.Logs, ibs.GetRawLogs(txTask.TxIndex)...)
-							return ret, err
-						}
-
-						chainReader := consensuschain.NewReader(pe.cfg.chainConfig, applyTx, pe.cfg.blockReader, pe.logger)
-						if pe.isMining {
-							_, _, err =
-								pe.cfg.engine.FinalizeAndAssemble(
-									pe.cfg.chainConfig, types.CopyHeader(txTask.Header), ibs, txTask.Txs, txTask.Uncles, blockReceipts,
-									txTask.Withdrawals, chainReader, syscall, nil, pe.logger)
-						} else {
-							_, err =
-								pe.cfg.engine.Finalize(
-									pe.cfg.chainConfig, types.CopyHeader(txTask.Header), ibs, txTask.Txs, txTask.Uncles, blockReceipts,
-									txTask.Withdrawals, chainReader, syscall, false, pe.logger)
-						}
-
-						if err != nil {
-							return state.StateUpdates{}, fmt.Errorf("can't finalize block: %w", err)
-						}
-
-						stateWriter := state.NewBufferedWriter(pe.rs, nil)
-
-						if err = ibs.MakeWriteSet(txTask.EvmBlockContext.Rules(txTask.Config), stateWriter); err != nil {
-							return state.StateUpdates{}, err
-						}
-
-						return stateWriter.WriteSet(), nil
-					}()
-
-					if err != nil {
-						return err
-					}
-
-					blockResult.ApplyCount += stateUpdates.UpdateCount()
-					if dbg.TraceApply && dbg.TraceBlock(blockResult.BlockNum) {
-						stateUpdates.TraceBlockUpdates(blockResult.BlockNum, true)
-						fmt.Println(blockResult.BlockNum, "apply count", blockResult.ApplyCount)
-					}
-
-					blockExecutor.applyResults <- &txResult{
-						blockNum:     blockResult.BlockNum,
-						txNum:        blockResult.lastTxNum,
-						rules:        result.Rules(),
-						stateUpdates: stateUpdates,
-						logs:         result.Logs,
-						traceFroms:   result.TraceFroms,
-						traceTos:     result.TraceTos,
-					}
-				}
-
-				if !blockExecutor.execStarted.IsZero() {
-					pe.blockExecMetrics.Duration.Add(time.Since(blockExecutor.execStarted))
-					pe.blockExecMetrics.BlockCount.Add(1)
-				}
-				blockExecutor.applyResults <- blockResult
-				pe.Lock()
-				delete(pe.blockExecutors, blockResult.BlockNum)
-				pe.Unlock()
-			}
-
-			pe.RLock()
-			blockExecutor, ok = pe.blockExecutors[blockResult.BlockNum+1]
-			pe.RUnlock()
-
-			if ok {
-				pe.onBlockStart(ctx, blockExecutor.blockNum, blockExecutor.blockHash)
-				blockExecutor.execStarted = time.Now()
-				blockExecutor.scheduleExecution(ctx, pe)
-			}
-		}
-	}
-}
-
-func (pe *parallelExecutor) processRequest(ctx context.Context, execRequest *execRequest) (err error) {
-	prevSenderTx := map[common.Address]int{}
-	var scheduleable *blockExecutor
-	var executor *blockExecutor
-
-	for i, txTask := range execRequest.tasks {
-		t := &execTask{
-			Task:               txTask,
-			index:              i,
-			shouldDelayFeeCalc: true,
-		}
-
-		blockNum := t.Version().BlockNum
-
-		if executor == nil {
-			var ok bool
-			executor, ok = pe.blockExecutors[blockNum]
-
-			if !ok {
-				executor = newBlockExec(blockNum, execRequest.blockHash, execRequest.gasPool, execRequest.applyResults, execRequest.profile)
-			}
-		}
-
-		executor.tasks = append(executor.tasks, t)
-		executor.results = append(executor.results, nil)
-		executor.txIncarnations = append(executor.txIncarnations, 0)
-		executor.execFailed = append(executor.execFailed, 0)
-		executor.execAborted = append(executor.execAborted, 0)
-
-		executor.skipCheck[len(executor.tasks)-1] = false
-		executor.estimateDeps[len(executor.tasks)-1] = []int{}
-
-		executor.execTasks.pushPending(i)
-		executor.validateTasks.pushPending(i)
-
-		if len(t.Dependencies()) > 0 {
-			for _, depTxIndex := range t.Dependencies() {
-				executor.execTasks.addDependency(depTxIndex+1, i)
-			}
-			executor.execTasks.clearPending(i)
-		} else {
-			sender, err := t.TxSender()
-			if err != nil {
-				return err
-			}
-			if sender != nil {
-				if tx, ok := prevSenderTx[*sender]; ok {
-					executor.execTasks.addDependency(tx, i)
-					executor.execTasks.clearPending(i)
-				}
-
-				prevSenderTx[*sender] = i
-			}
-		}
-
-		if t.IsBlockEnd() {
-			pe.Lock()
-			if len(pe.blockExecutors) == 0 {
-				pe.blockExecutors = map[uint64]*blockExecutor{
-					blockNum: executor,
-				}
-				scheduleable = executor
-			} else {
-				pe.blockExecutors[t.Version().BlockNum] = executor
-			}
-			pe.Unlock()
-
-			executor = nil
-		}
-	}
-
-	if scheduleable != nil {
-		pe.blockExecMetrics.BlockCount.Add(1)
-		scheduleable.execStarted = time.Now()
-		scheduleable.scheduleExecution(ctx, pe)
-	}
-
-	return nil
-}
-
-func (pe *parallelExecutor) processResults(ctx context.Context, applyTx kv.TemporalTx) (blockResult *blockResult, err error) {
-	rwsIt := pe.rws.Iter()
-	for rwsIt.HasNext() && (blockResult == nil || !blockResult.complete) {
-		txResult := rwsIt.PopNext()
-
-		if pe.cfg.syncCfg.ChaosMonkey && pe.enableChaosMonkey {
-			chaosErr := chaos_monkey.ThrowRandomConsensusError(false, txResult.Version().TxIndex, pe.cfg.badBlockHalt, txResult.Err)
-			if chaosErr != nil {
-				log.Warn("Monkey in consensus")
-				return blockResult, chaosErr
-			}
-		}
-
-		pe.RLock()
-		blockExecutor, ok := pe.blockExecutors[txResult.Version().BlockNum]
-		pe.RUnlock()
-
-		if !ok {
-			return nil, fmt.Errorf("unknown block: %d", txResult.Version().BlockNum)
-		}
-
-		blockResult, err = blockExecutor.nextResult(ctx, pe, txResult, applyTx)
-
-		if err != nil {
-			return blockResult, err
-		}
-	}
-
-	return blockResult, nil
-}
-
-func (pe *parallelExecutor) run(ctx context.Context) (context.Context, context.CancelFunc) {
-	pe.execRequests = make(chan *execRequest, 100_000)
-	pe.in = exec.NewQueueWithRetry(100_000)
-
-	pe.taskExecMetrics = exec3.NewWorkerMetrics()
-	pe.blockExecMetrics = newBlockExecMetrics()
-
-	execLoopCtx, execLoopCtxCancel := context.WithCancel(ctx)
-	pe.execLoopGroup, execLoopCtx = errgroup.WithContext(execLoopCtx)
-
-	pe.execWorkers, _, pe.rws, pe.stopWorkers, pe.waitWorkers = exec3.NewWorkersPool(
-		execLoopCtx, nil, true, pe.cfg.db, nil, nil, nil, pe.in,
-		pe.cfg.blockReader, pe.cfg.chainConfig, pe.cfg.genesis, pe.cfg.engine,
-		pe.workerCount+1, pe.taskExecMetrics, pe.cfg.dirs, pe.isMining, pe.logger)
-
-	pe.execLoopGroup.Go(func() error {
-		defer pe.rws.Close()
-		defer pe.in.Close()
-		pe.resetWorkers(execLoopCtx, pe.rs, nil)
-		return pe.execLoop(execLoopCtx)
-	})
-
-	return execLoopCtx, func() {
-		execLoopCtxCancel()
-		pe.wait(ctx)
-		pe.stopWorkers()
-		pe.in.Close()
-	}
-}
-
-func (pe *parallelExecutor) wait(ctx context.Context) error {
-	doneCh := make(chan error, 1)
-
-	go func() {
-		if pe.execLoopGroup != nil {
-			err := pe.execLoopGroup.Wait()
-			if err != nil && !errors.Is(err, context.Canceled) {
-				doneCh <- err
-				return
-			}
-			pe.waitWorkers()
-		}
-		doneCh <- nil
-	}()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return nil
-		case err := <-doneCh:
-			return err
 		}
 	}
 }
